### 2024-03-17 #633 @jmaspons
osmapiR: an implementation of OpenStreetMap API v0.6 for R 
Submitting Author Name: Joan Maspons
Submitting Author Github Handle: @jmaspons
Repository:  https://github.com/jmaspons/osmapiR
Version submitted: 0.0.0.22
Submission type: Standard
Version accepted: TBD

---

```
Package: osmapiR
Title: OpenStreetMap API
Version: 0.0.0.22
    person("Joan", "Maspons", , "joanmaspons@gmail.com", role = c("aut", "cre"),
Maintainer: Joan Maspons <joanmaspons@gmail.com>
Description: Implements OpenStreetMap API v0.6
    (<https://wiki.openstreetmap.org/wiki/API_v0.6>).
    curl,
    httr2,
    xml2
    httptest2,
    httpuv,
X-schema.org-keywords: open street map, openstreetmap, OSM, openstreetmap-api, osmapi, API
```

`osmapiR` allows retrieving all kind of data from the OpenStreetMap project.

Useful for people working in OSM and to study the evolution of the map.

`osmapiR` is the only package to access other OpenStreetMap data than the maps data (map notes, GPS traces, changelogs and users). It can be also useful to get the history of the OSM objects and is the only package that allows editing and upload any kind of data.

To access the OSM map data for purposes other than editing or exploring the history of the objects, perhaps is better to use the other packages that implements the Overpass API ([osmdata](https://github.com/ropensci/osmdata)) or that works with .pbf files ([osmexcract](https://github.com/ropensci/osmextract/)).

The package allows accessing all the public data from www.openstreetmap.org. All calls use https and OAuth2 is used for authenticated calls. I redirect the user to osmdata or osmecract in the README and in the documentation of the functions to get map data as recommended by the [OSM policies](https://operations.osmfoundation.org/policies/api/).

Some functions implement API calls restricted to users with moderator role and are not well documented in the OSM wiki. These functions lack a return value in docs.  Some functions that cause editions in the OSM servers don't have examples to avoid that the user accidentally run it.
