### 2024-03-18 jmaspons/osmapiR#13 @mpadge
And BTW, installing the [`pkgcheck-action`](https://github.com/ropensci-review-tools/pkgcheck-action) in this repo would have helped diagnose that. I'm sure it will help you progress through review if you have that workflow running here too.
