### 2024-03-18 7adc756
Remove explicit doc links to osmdata pkg (not a dependence) FIX #13
