+++ R/osmapi_elements.R
@@ -679,7 +679,7 @@ osm_version_object <- function(osm_type = c("node", "way", "relation"), osm_id, 
 #' @note
 #' For downloading data for purposes other than editing or exploring the history of the objects, perhaps is better to
 #' use the Overpass API. A similar function to download OSM objects by `type` and `id` using Overpass, is implemented in
-#' [osmdata::opq_osm_id()].
+#' the \pkg{osmdata} function `opq_osm_id()`.
 #'
 #' @return
 #' If `format = "R"`, returns a data frame with one OSM object per row. If `format = "xml"`, returns a
@@ -906,7 +906,7 @@ osm_ways_node <- function(node_id, format = c("R", "xml", "json"), tags_in_colum
 #' @note
 #' For downloading data for purposes other than editing or exploring the history of the objects, perhaps is better to
 #' use the Overpass API. A similar function to download OSM objects by `type` and `id` using Overpass, is implemented in
-#' [osmdata::opq_osm_id()].
+#' the \pkg{osmdata} function `opq_osm_id()`.
 #'
 #' @return
 #' If `format = "R"`, returns a data frame with one OSM object per row. If `format = "xml"`, returns a
