### 2024-03-18 jmaspons/osmapiR#13 @jmaspons
Thanks, fixed! I will have a look to the pkgcheck-action
