### 2024-03-18 a126bde
pkgcheck::use_github_action_pkgcheck()
