### 2024-03-19 9f01d35
Update docs following wiki page up to v2681190
https://wiki.openstreetmap.org/w/index.php?title=API_v0.6&oldid=2681190
https://wiki.openstreetmap.org/w/index.php?title=API_v0.6&diff=cur&oldid=2681190
