### 2024-03-19 5878c08
Test osm_query_changesets(..., order = "oldest")
