### 2024-03-21 4e0c463
Parse osm_delete_note() response, add text parameter and test it
