### 2024-03-21 9e55b7f
Fix osm_redaction_object and add @return. Tests fail
