### 2024-03-21 40ad1aa
Test and document osm_[un]hide_comment_changeset_discussion()
