+++ tests/testthat/test-miscellaneous.R
@@ -51,7 +51,7 @@ test_that("osm_bbox_objects works", {
 
 
   # methods
-  expect_s3_class(print(bbox_objects), c("osmapi_objects", "data.frame"))
+  expect_s3_class(print(bbox_objects), c("osmapi_objects", "data.frame"), exact = TRUE)
 
 
   ## Empty results
@@ -60,7 +60,7 @@ test_that("osm_bbox_objects works", {
     empty_bbox_objects <- osm_bbox_objects(bbox = c(-180, 0, -179.9, 0.1))
   })
 
-  expect_s3_class(empty_bbox_objects, c("osmapi_objects", "data.frame"))
+  expect_s3_class(empty_bbox_objects, c("osmapi_objects", "data.frame"), exact = TRUE)
   expect_identical(names(empty_bbox_objects), obj_cols)
   expect_named(attr(empty_bbox_objects, "bbox"), c("minlat", "minlon", "maxlat", "maxlon"))
   expect_identical(nrow(empty_bbox_objects), 0L)
@@ -73,7 +73,7 @@ test_that("osm_bbox_objects works", {
 
 
   # methods
-  expect_s3_class(print(empty_bbox_objects), c("osmapi_objects", "data.frame"))
+  expect_s3_class(print(empty_bbox_objects), c("osmapi_objects", "data.frame"), exact = TRUE)
 })
 
 
