### 2024-03-21 #633 @jmaspons
All returns are now documented. Some requests return nothing, others an xml that is parsed to a data.frame + package classes
