### 2024-03-21 78f43fa
Make user_agent modifiable and remember API Usage Policy
Requested by pnorman at irc.oftc.net #osm-dev
