+++ R/osmapiR-package.R
@@ -1,3 +1,12 @@
+#' @details
+#' You are responsible for following the [API Usage Policy](https://operations.osmfoundation.org/policies/api/).
+#'
+#' You can modify the user agent of the requests by setting the option `osmapir.user_agent`:
+#' ```r
+#' options(osmapir.user_agent = "my new user agent")
+#' ```
+#'
+#'
 #' @keywords internal
 "_PACKAGE"
 
