+++ tests/testthat/setup.R
@@ -2,4 +2,5 @@ library(httptest2)
 
 set_osmapi_connection("testing")
 options(osmapir.osmapir_version = "")
+options(osmapir.user_agent = "osmapiR (https://github.com/jmaspons/osmapiR)")
 options(osmapir.R_CMD_check = TRUE)
