### 2024-03-21 b4a594b
Improve documentation of the package
From _pkgdown and https://jmaspons.github.io/osmapiR/reference/index.html
