+++ R/osmapiR-package.R
@@ -6,6 +6,211 @@
 #' options(osmapir.user_agent = "my new user agent")
 #' ```
 #'
+#' # Overview of the function
+#' All function starting with `osm_*` include calls to the server.
+#'
+#' ## OSM objects
+#'
+#' ### Get OSM objects
+#'
+#' [osm_bbox_objects()]
+#'     Retrieve map data by bounding box
+#'
+#' [osm_get_objects()]
+#'     Get OSM objects
+#'
+#' [osm_history_object()]
+#'     Get the history of an object
+#'
+#' [osm_relations_object()]
+#'     Relations of an object
+#'
+#' [osm_ways_node()]
+#'     Ways of a node
+#'
+#' [osmapi_objects()]
+#'     osmapi_objects
+#'
+#' ### Edit OSM objects
+#'
+#' [osm_create_object()]
+#'     Create an OSM object
+#'
+#' [osm_delete_object()]
+#'     Delete an OSM object
+#'
+#' [osm_update_object()]
+#'     Update an OSM object
+#'
+#' ## Changesets
+#'
+#' Every modification of the standard OSM elements has to reference an open changeset. A changeset may contain tags
+#' just like the other elements. A recommended tag for changesets is the key ''comment=*'' with a short human readable
+#' description of the changes being made in that changeset. A new changeset can be opened at any time and a changeset
+#' may be referenced from multiple API calls. Because of this it can be closed manually as the server can't know when
+#' one changeset ends and another should begin. To avoid stale open changesets a mechanism is implemented to
+#' automatically close changesets. See [OSM wiki](https://wiki.openstreetmap.org/wiki/Changeset) for details.
+#'
+#' ### Get changesets
+#'
+#' [osm_download_changeset()]
+#'     Download a changeset in OsmChange format
+#'
+#' [osm_get_changesets()]
+#'     Get changesets
+#'
+#' [osm_query_changesets()]
+#'     Query changesets
+#'
+#' ### Edit changeset
+#'
+#' [osm_create_changeset()] [osm_update_changeset()] [osm_close_changeset()]
+#'     Create, update, or close a changeset
+#'
+#' [osm_diff_upload_changeset()]
+#'     Diff (OsmChange format) upload to a changeset
+#'
+#' ### Changeset's discussion"
+#'
+#' [osm_comment_changeset_discussion()]
+#'     Comment a changeset
+#'
+#' [osm_hide_comment_changeset_discussion()] [osm_unhide_comment_changeset_discussion()]
+#'     Hide or unhide a changeset comment
+#'
+#' [osm_subscribe_changeset_discussion()] [osm_unsubscribe_changeset_discussion()]
+#'     Subscribe or unsubscribe to a changeset discussion
+#'
+#' ## Map notes
+#'
+#' This functions provides access to the [notes](https://wiki.openstreetmap.org/wiki/Notes) feature, which allows users
+#' to add geo-referenced textual \"post-it\" notes.
+#'
+#' ### Get map notes
+#'
+#' [osm_feed_notes()]
+#'     RSS Feed of notes in a bbox
+#'
+#' [osm_get_notes()]
+#'     Get notes
+#'
+#' [osm_read_bbox_notes()]
+#'     Retrieve notes by bounding box
+#'
+#' [osm_search_notes()]
+#'     Search for notes
+#'
+#' ### Edit map notes
+#'
+#' osm_close_note()] [osm_reopen_note()]
+#'     Close or reopen a note
+#'
+#' [osm_create_comment_note()]
+#'     Create a new comment in a note
+#'
+#' [osm_create_note()]
+#'     Create a new note
[…]
