+++ man/osmapiR-package.Rd
@@ -16,6 +16,229 @@ You can modify the user agent of the requests by setting the option \code{osmapi
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{options(osmapir.user_agent = "my new user agent")
 }\if{html}{\out{</div>}}
 }
+\section{Overview of the function}{
+All function starting with \verb{osm_*} include calls to the server.
+\subsection{OSM objects}{
+\subsection{Get OSM objects}{
+
+\code{\link[=osm_bbox_objects]{osm_bbox_objects()}}
+Retrieve map data by bounding box
+
+\code{\link[=osm_get_objects]{osm_get_objects()}}
+Get OSM objects
+
+\code{\link[=osm_history_object]{osm_history_object()}}
+Get the history of an object
+
+\code{\link[=osm_relations_object]{osm_relations_object()}}
+Relations of an object
+
+\code{\link[=osm_ways_node]{osm_ways_node()}}
+Ways of a node
+
+\code{\link[=osmapi_objects]{osmapi_objects()}}
+osmapi_objects
+}
+
+\subsection{Edit OSM objects}{
+
+\code{\link[=osm_create_object]{osm_create_object()}}
+Create an OSM object
+
+\code{\link[=osm_delete_object]{osm_delete_object()}}
+Delete an OSM object
+
+\code{\link[=osm_update_object]{osm_update_object()}}
+Update an OSM object
+}
+
+}
+
+\subsection{Changesets}{
+
+Every modification of the standard OSM elements has to reference an open changeset. A changeset may contain tags
+just like the other elements. A recommended tag for changesets is the key ''comment=*'' with a short human readable
+description of the changes being made in that changeset. A new changeset can be opened at any time and a changeset
+may be referenced from multiple API calls. Because of this it can be closed manually as the server can't know when
+one changeset ends and another should begin. To avoid stale open changesets a mechanism is implemented to
+automatically close changesets. See \href{https://wiki.openstreetmap.org/wiki/Changeset}{OSM wiki} for details.
+\subsection{Get changesets}{
+
+\code{\link[=osm_download_changeset]{osm_download_changeset()}}
+Download a changeset in OsmChange format
+
+\code{\link[=osm_get_changesets]{osm_get_changesets()}}
+Get changesets
+
+\code{\link[=osm_query_changesets]{osm_query_changesets()}}
+Query changesets
+}
+
+\subsection{Edit changeset}{
+
+\code{\link[=osm_create_changeset]{osm_create_changeset()}} \code{\link[=osm_update_changeset]{osm_update_changeset()}} \code{\link[=osm_close_changeset]{osm_close_changeset()}}
+Create, update, or close a changeset
+
+\code{\link[=osm_diff_upload_changeset]{osm_diff_upload_changeset()}}
+Diff (OsmChange format) upload to a changeset
+}
+
+\subsection{Changeset's discussion"}{
+
+\code{\link[=osm_comment_changeset_discussion]{osm_comment_changeset_discussion()}}
+Comment a changeset
+
+\code{\link[=osm_hide_comment_changeset_discussion]{osm_hide_comment_changeset_discussion()}} \code{\link[=osm_unhide_comment_changeset_discussion]{osm_unhide_comment_changeset_discussion()}}
+Hide or unhide a changeset comment
+
[…]
