### 2024-03-21 ca68fbc
Avoid using sapply. Using vapply instead, that's type safe
