### 2024-03-23 99a34e2
Document changing user_agent and remember API Usage Policy
Requested by pnorman at irc.oftc.net #osm-dev
