+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# osmapiR
+# osmapiR <a href="https://jmaspons.github.io/osmapiR"><img src="man/figures/logo.svg" align="right" height="139" alt="osmapiR website" /></a>
 
 <!-- badges: start -->
 
