+++ README.Rmd
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# osmapiR
+# osmapiR <a href="https://jmaspons.github.io/osmapiR"><img src="man/figures/logo.svg" align="right" height="139" alt="osmapiR website" /></a>
 
 <!-- badges: start -->
 [![R-CMD-check](https://github.com/jmaspons/osmapiR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jmaspons/osmapiR/actions/workflows/R-CMD-check.yaml)
