### 2024-03-23 ca46c42
Increase version and logo size
