### 2024-03-25 cc2ff89
Fix error + tests + remove some TODOs fixed in the server side for gps traces
