### 2024-03-26 c25b53a
Implement and test osm_set_preferences_user()
