### 2024-04-04 0fe330c
Redact headers in mock files for tests while resolving #16
