### 2024-04-04 d61c24a
Add parameter keep_na_tags at osmapi_objects
