### 2024-04-18 #633 @adamhsparks
### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via GitHub Actions
- [x] **License:** The package has a CRAN accepted license

---

#### Editor comments

Hi there.

I've found a few issues flagged by the initial editor checks:

##### Missing @examples

You have a few empty `#' @examples` tags that need to be provided

```r
✖ osmapi_gps_traces.R:180: @examples requires a value.
✖ osmapi_gps_traces.R:222: @examples requires a value.
✖ osmapi_gps_traces.R:250: @examples requires a value.
```

#### [`goodpractice::gp`](https://github.com/MangoTheCat/goodpractice/) output

```
── GP osmapiR ───────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 86% of code lines are
    covered by test cases.

    R/osm_get_changesets.R:97:NA
    R/osm_get_changesets.R:98:NA
    R/osm_get_changesets.R:99:NA
    R/osm_get_objects.R:65:NA
    R/osm_get_objects.R:68:NA
    ... and 251 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines shorter than 80 characters

    inst/httptest2/redact.R:19:81
    inst/httptest2/redact.R:20:81
    inst/httptest2/redact.R:31:81
    R/oauth2.R:28:81
    R/oauth2.R:29:81
    ... and 1330 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the
    input data. Consider using vapply() instead.

    tests/testthat/test-osmapiR_objects.R:60:21

─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

**Reviewers** note highlighted coverage gaps and please inspect what is not covered to see if you think important areas are missed.

#### `devtools::spell_check()`

You might consider adding the 'Language' field to your DESCRIPTION file.

```
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
No spelling errors found.
```

#### Comments on Documentation

I noted the use of `''`, I would use `dQuote{}`, myself, and `<code>`, here I would use ` `` ` to denote code blocks, among other things in the ROxygen headers.
I would suggest reviewing the ROxygen documentation <https://roxygen2.r-lib.org/index.html> and possibly enabling Markdown support for formatting the Rd output.

The README has a misspelled R Package name, it should be ({osmextract})[https://github.com/ropensci/osmextract].
I'd also suggest linking to the other OSM packages, _e.g._ {osmdata}, referred to in the README rather than just leaving them as plain text.

─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

Happy to start looking for reviewers. I recommend using this time to try and address some of the issues raised so they can be checked during review rather that be reflagged.

Let me know if anything is unclear!
