### 2024-04-22 3fe7911
sapply -> vapply
