+++ README.Rmd
@@ -21,6 +21,7 @@ knitr::opts_chunk$set(
 [![pkgdown](https://github.com/jmaspons/osmapiR/actions/workflows/pkgdown.yaml/badge.svg)](https://github.com/jmaspons/osmapiR/actions/workflows/pkgdown.yaml)
 [![Project Status: Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/osmapiR)](https://cran.r-project.org/package=osmapiR)
+[![](https://badges.ropensci.org/633_status.svg)](https://github.com/ropensci/software-review/issues/633)
 <!-- badges: end -->
 
 Implements [OpenStreetMap API](https://wiki.openstreetmap.org/wiki/API_v0.6) for R.
@@ -93,7 +94,7 @@ gpx
 ## Related packages
 
 * [osmdata](https://cran.r-project.org/package=osmdata) implements the Overpass API to query data from OSM.
-* [osmexctract](https://cran.r-project.org/package=osmextract) matches, downloads, converts and imports bulk OSM data
+* [osmextract](https://cran.r-project.org/package=osmextract) matches, downloads, converts and imports bulk OSM data
   (`.pbf` files)
 * [OpenStreetMap](https://cran.r-project.org/package=OpenStreetMap) Accesses high resolution raster maps using the
   OpenStreetMap protocol.
@@ -103,7 +104,8 @@ users). It can be also useful to get the history of the OSM objects and is the o
 upload any kind of data.
 
 To access the OSM map data for purposes other than editing or exploring the history of the objects, perhaps is better to
-use the other packages that implements the Overpass API (`osmdata`) or that works with `.pbf` files (`osmexcract`).
+use the other packages that implements the Overpass API ([osmdata](https://cran.r-project.org/package=osmdata)) or
+that works with `.pbf` files ([osmextract](https://cran.r-project.org/package=osmextract)).
 
 ## Code of Conduct
 
