### 2024-04-25 2baf223
Redact headers and refresh mock files for tests
TODO: request headers not redacted (e.g. Authorization)
