+++ inst/httptest2/redact.R
@@ -28,7 +28,10 @@ function(response) {
   )
 
   # simplify = FALSE
-  response <- httptest2::redact_headers(response, headers = c("x-request-id", "date", "Authorization"))
+  response <- httptest2::redact_headers(
+    response,
+    headers = c("x-request-id", "date", "Authorization", "x-amz-id-2", "x-amz-request-id")
+  )
 
   return(response)
 }
