+++ tests/testthat/mock_get_data_gpx/osm.org/api/0.6/gpx/3458743/data.gpx.R
@@ -2,12 +2,12 @@ structure(list(method = "GET", url = "osm.org/api/0.6/gpx/3458743/data.gpx",
     status_code = 200L, headers = structure(list(`cache-control` = "max-age=0, private, must-revalidate", 
         vary = "Accept-Language,Origin", `referrer-policy` = "strict-origin-when-cross-origin", 
         `x-permitted-cross-domain-policies` = "none", `x-xss-protection` = "1; mode=block", 
-        `content-language` = "en", `x-request-id` = "ZgFIOlQpm25QfdliuMAdhgACIw4", 
+        `content-language` = "en", `x-request-id` = "REDACTED", 
         `content-disposition` = "attachment; filename=\"3458743.gpx\"; filename*=UTF-8''3458743.gpx", 
         `x-download-options` = "noopen", `content-transfer-encoding` = "binary", 
         `x-runtime` = "0.252595", `x-frame-options` = "sameorigin", 
         `x-content-type-options` = "nosniff", `content-security-policy` = "default-src 'self'; child-src 'self'; connect-src 'self' matomo.openstreetmap.org; font-src 'none'; form-action 'self'; frame-ancestors 'self'; frame-src 'self'; img-src 'self' data: www.gravatar.com *.wp.com tile.openstreetmap.org *.tile.openstreetmap.org *.tile.thunderforest.com tile.tracestrack.com *.openstreetmap.fr matomo.openstreetmap.org https://openstreetmap-user-avatars.s3.dualstack.eu-west-1.amazonaws.com https://openstreetmap-gps-images.s3.dualstack.eu-west-1.amazonaws.com; manifest-src 'self'; media-src 'none'; object-src 'self'; script-src 'self' matomo.openstreetmap.org; style-src 'self'; worker-src 'none'", 
-        date = "Mon, 25 Mar 2024 09:47:38 GMT", `x-powered-by` = "Phusion Passenger(R) 6.0.20", 
+        date = "REDACTED", `x-powered-by` = "Phusion Passenger(R) 6.0.20", 
         `strict-transport-security` = "max-age=31536000; includeSubDomains; preload", 
         etag = "W/\"08e602c081d848ede98c2ac68bf30ed0\"", status = "200 OK", 
         `content-type` = "application/gpx+xml", server = "Apache/2.4.54 (Ubuntu)"), class = "httr2_headers"), 
@@ -6266,7 +6266,7 @@ structure(list(method = "GET", url = "osm.org/api/0.6/gpx/3458743/data.gpx",
     0x2f, 0x67, 0x70, 0x78, 0x3e)), request = structure(list(
         url = "https://api.openstreetmap.org/api/0.6/gpx/3458743/data.gpx", 
         method = "GET", headers = structure(list(Authorization = "Bearer kcV3MYRNTBJOt-1Tqd5Yfv-jaAUEbDcbmynR0TdZ5ek"), redact = "Authorization"), 
-        body = NULL, fields = list(), options = list(useragent = "osmapiR 0.0.0.25 (https://github.com/jmaspons/osmapiR)"), 
+        body = NULL, fields = list(), options = list(useragent = "osmapiR (https://github.com/jmaspons/osmapiR)"), 
         policies = list(retry_max_tries = 10L, error_body = function (resp) 
         {
             out <- switch(httr2::resp_content_type(resp), `text/plain` = httr2::resp_body_string(resp), 
