### 2024-04-26 f05bec7
Fix osm_update_gpx()
Tests still failing with httptest2
