### 2024-04-26 91a6ddb
Test osm_diff_upload_changeset()
