### 2024-04-28 b817d32
Improve test coverage and fix errors found
