+++ tests/testthat/mock_error_responses/osm.org/api/0.6/way/1.R
@@ -1,0 +1,6 @@
+structure(list(method = "GET", url = "osm.org/api/0.6/way/1", 
+    status_code = 404L, headers = structure(list(`cache-control` = "no-cache", 
+        `strict-transport-security` = "max-age=31536000; includeSubDomains; preload", 
+        `content-length` = "0", `content-type` = "text/html; charset=utf-8", 
+        date = "REDACTED", server = "Apache/2.4.54 (Ubuntu)"), class = "httr2_headers"), 
+    body = raw(0), cache = new.env(parent = emptyenv())), class = "httr2_response")
