### 2024-04-28 b51e9dc
Tests for error responses with html body
