### 2024-04-29 9c10c9f
Improve test coverage of functions to edit gps traces
Requires https://github.com/nealrichardson/httptest2/pull/40
