### 2024-04-29 bb79ad1
Deploying to gh-pages from @ jmaspons/osmapiR@9c10c9f673bf5e77a2eabf0498004798b28a1669 🚀
