### 2024-04-29 26a7b47
Revert "Improve test coverage of functions to edit gps traces"
This reverts commit 9c10c9f673bf5e77a2eabf0498004798b28a1669 which
requires https://github.com/nealrichardson/httptest2/pull/40
