### 2024-04-29 5c4ddc5
Revert "Revert "Improve test coverage of functions to edit gps traces""
This reverts commit 26a7b473062a917ec21e29c859a9797db79c1db2.
