### 2024-05-17 e3984fb
Avoid HTTP ERROR 414: URI Too Long in osm_fetch_objects() by calls in batches
Maximize batch size to URIs <= 8213 characters
