### 2024-05-17 92c30fe
Deactivate tests that work when running interactively but fail in R CMD check
