+++ tests/testthat/test-elements.R
@@ -281,116 +281,123 @@ test_that("osm_fetch_objects works", {
 
   ### Test long URL in batches to avoid ERROR: HTTP 414 URI Too Long ----
 
-  # osm_ids <- unique(sort(toponimsCat::municipis$id))
-  # seq_ids <- list(osm_ids[1])
-  # k <- 1
-  # for (i in seq_len(length(osm_ids))[-1]) {
-  #   if (osm_ids[i] - osm_ids[i - 1] > 1) {
-  #     k <- k + 1
-  #     seq_ids[[k]] <- osm_ids[i]
-  #   } else if (osm_ids[i + 1] - osm_ids[i] > 1) {
-  #     seq_ids[[k]] <- c(seq_ids[[k]], osm_ids[i])
-  #   }
-  # }
-  # cmd <- paste0("osm_ids <- c(", paste(sapply(seq_ids, function(x) paste(x, collapse = ":")), collapse = ", "), ")")
-  # all.equal(osm_ids, eval(parse(text = cmd)))
-  # cat(cmd)
+  # Tests work when running interactively but fail in R CMD check:
+  #
+  #   Error in `stop_request(req)`: An unexpected request was made:
+  #   GET https://master.apis.dev.openstreetmap.org/api/0.6/relations?relations=...
+  #
+  #   Expected mock file: osm.org/api/0.6/relations-13eee2.*
 
-  osm_ids <- c( ## Municipis PPCC
-    18000, 18311, 18316, 18318, 18326, 18328, 18349:18352, 18354, 18362:18363, 18375, 18391, 18409, 18416:18417, 18419,
-    18428, 18475:18479, 18482, 18484, 18496, 20224, 20339, 22531, 22588, 23234, 23237, 23266, 23303, 23308, 23315,
-    23324, 23340, 23759, 23763, 23790, 23795, 23804, 23814, 23823, 23895, 24457:24459, 24858, 24940, 25873, 27529,
-    34167, 40581, 51547, 53365, 54461, 54467, 54469, 54471, 74277, 74281, 74284, 74308:74310, 74988, 74993, 80067,
-    81599, 123160, 269776, 270405, 271192, 271356, 271429, 271495, 271536, 271664, 271713, 272157, 273497:273498,
-    274180, 339488:339490, 339492:339493, 339495, 339498:339503, 339506:339511, 339513:339515, 339534, 339536:339537,
-    339568, 339572:339574, 339577, 339584, 339613:339622, 339664, 339809:339811, 339815:339821, 339823:339824, 339847,
-    339849:339852, 339854:339855, 339858, 339862, 339873:339880, 339887, 339928, 339930, 339963:339972, 340014, 340016,
-    340030, 340174:340179, 340193:340204, 340206:340211, 340216:340217, 340257, 340259, 340261, 340307, 340319,
-    340328:340331, 340341:340351, 340373:340375, 340377, 340379:340382, 340390, 340413, 340438:340439, 340441:340442,
-    340445:340448, 340454, 340476, 340485, 340487:340488, 340492, 340498, 340500:340502, 340506:340507, 340518,
-    340526:340528, 340530:340553, 340555, 340561:340562, 340661, 340689, 340694, 340696:340697, 340700, 340791, 340826,
-    340828, 340843, 340848:340856, 340858:340859, 340865:340872, 340908, 340920, 340928, 340930:340933, 340935, 340975,
-    341046:341047, 341049:341051, 341053, 341056, 341135, 341140:341142, 341144:341148, 341193, 341195, 341224,
-    341230:341236, 341240:341241, 341308:341311, 341314:341315, 341318:341321, 341323, 341330, 341355, 341395:341397,
[…]
