### 2024-05-17 c50fdb1
Warnings for ignored object editions (more than one element)
Point to osmchange_*() + osm_diff_upload_changeset()
