### 2024-05-17 bb81574
Improve parsing of server errors
