+++ tests/testthat/test-osmapiR_request.R
@@ -23,11 +23,19 @@ test_that("error handling works", {
       "The user doesn't own that changeset"
     )
 
+
     ##  Message in the body (content-type: text/html)
+
     expect_error(
       osmapi_request() |> httr2::req_url_path_append("err") |> httr2::req_perform(),
-      "File not found"
+      "404 File not found"
     )
-    # expect_error(osm_delete_gpx(1:2)) may be fixed in the future by vectorizing api calls
+    # expect_error(osm_delete_gpx(1:2)) # may be fixed in the future by vectorizing api calls
+
+    expect_error(
+      osmapi_request() |> httr2::req_url_path_append("nodes") |>
+        httr2::req_url_query(nodes = paste(1:2000, collapse = ",")) |> httr2::req_perform(),
+      "414 URI Too Long"
+    )
   })
 })
