### 2024-05-17 837a85c
Link docs to xml2::xml_document-class
