### 2024-05-17 ae7af03
Add  parameter to osmchange_*() to return  in osc/xml format
