+++ R/osmchange.R
@@ -8,13 +8,20 @@
 #'   all tags will be updated, removed or created. If `FALSE`, don't modify tags.
 #' @param members If `TRUE` and `x` has a `members` column, update the members of the ways and relations objects.
 #' @param lat_lon If `TRUE` and `x` has a `lat` and `lon` columns, update the coordinates of the node objects.
+#' @param format Format of the output. Can be `R` (default), `osc` (`xml` is a synonym for `osc`).
 #'
 #' @details
 #' `x` should be a `osmapi_objects` or follow the same format. Missing tags or tags with `NA` in the value will be
 #' removed if `tag_keys` is not specified. See [osm_get_objects()] for examples of the format.
 #'
-#' @return Returns a `osmapi_OsmChange` data frame with one OSM object per row ready to send the editions to the servers
-#'   with [osm_diff_upload_changeset()].
+#' @return
+#' If `format = "R"`, returns a `osmapi_OsmChange` data frame with one OSM edition per row.
+#' If `format = "osc"` or `format = "xml"`, returns a [xml2::xml_document-class] following the
+#' [OsmChange format](https://wiki.openstreetmap.org/wiki/OsmChange) that can be saved with [xml2::write_xml()] and
+#' opened in other applications such as JOSM.
+#'
+#' The results are  ready to send the editions to the servers with [osm_diff_upload_changeset()].
+#'
 #' @family OsmChange's functions
 #' @export
 #'
@@ -28,7 +35,8 @@
 #' osmch <- osmchange_modify(obj)
 #' osmch
 #' }
-osmchange_modify <- function(x, tag_keys, members = FALSE, lat_lon = FALSE) {
+osmchange_modify <- function(x, tag_keys, members = FALSE, lat_lon = FALSE, format = c("R", "osc", "xml")) {
+  format <- match.arg(format)
   stopifnot(inherits(x, "osmapi_objects"))
   if (inherits(x, "tags_wide")) {
     x <- tags_wide2list(x)
@@ -98,7 +106,11 @@ osmchange_modify <- function(x, tag_keys, members = FALSE, lat_lon = FALSE) {
     osmchange <- osmchange[!rm, ]
   }
 
-  class(osmchange) <- c("osmapi_OsmChange", "osmapi_objects", "data.frame")
+  if (format == "R") {
+    class(osmchange) <- c("osmapi_OsmChange", "osmapi_objects", "data.frame")
+  } else {
+    osmchange <- osmcha_DF2xml(osmchange)
+  }
 
   return(osmchange)
 }
@@ -112,6 +124,7 @@ osmchange_modify <- function(x, tag_keys, members = FALSE, lat_lon = FALSE) {
 #'   will be ignored.
 #' @param delete_if_unused If `TRUE`, the `if-unused` attribute will be added (see details). Can be a vector of length
 #'   `nrow(x)`.
+#' @param format Format of the output. Can be `R` (default), `osc` (`xml` is a synonym for `osc`).
 #'
 #' @details
 #' If `if-unused` attribute is present, then the delete operation(s) in this block are conditional and will only be
@@ -119,8 +132,14 @@ osmchange_modify <- function(x, tag_keys, members = FALSE, lat_lon = FALSE) {
 #' lead to an error, and the whole diff upload would fail. Setting the attribute will also cause deletions of already
 #' deleted objects to not generate an error.
 #'
-#' @return Returns a `osmapi_OsmChange` data frame with one OSM object per row ready to send the editions to the servers
-#'   with [osm_diff_upload_changeset()].
+#' @return
[…]
