+++ man/osmchange_modify.Rd
@@ -4,7 +4,13 @@
 \alias{osmchange_modify}
 \title{\code{osmchange} to modify existing OSM objects}
 \usage{
-osmchange_modify(x, tag_keys, members = FALSE, lat_lon = FALSE)
+osmchange_modify(
+  x,
+  tag_keys,
+  members = FALSE,
+  lat_lon = FALSE,
+  format = c("R", "osc", "xml")
+)
 }
 \arguments{
 \item{x}{A \link{osmapi_objects} with the columns \code{type} and \code{id} with unique combinations of values plus columns
@@ -16,10 +22,16 @@ all tags will be updated, removed or created. If \code{FALSE}, don't modify tags
 \item{members}{If \code{TRUE} and \code{x} has a \code{members} column, update the members of the ways and relations objects.}
 
 \item{lat_lon}{If \code{TRUE} and \code{x} has a \code{lat} and \code{lon} columns, update the coordinates of the node objects.}
+
+\item{format}{Format of the output. Can be \code{R} (default), \code{osc} (\code{xml} is a synonym for \code{osc}).}
 }
 \value{
-Returns a \code{osmapi_OsmChange} data frame with one OSM object per row ready to send the editions to the servers
-with \code{\link[=osm_diff_upload_changeset]{osm_diff_upload_changeset()}}.
+If \code{format = "R"}, returns a \code{osmapi_OsmChange} data frame with one OSM edition per row.
+If \code{format = "osc"} or \code{format = "xml"}, returns a \link[xml2:oldclass]{xml2::xml_document} following the
+\href{https://wiki.openstreetmap.org/wiki/OsmChange}{OsmChange format} that can be saved with \code{\link[xml2:write_xml]{xml2::write_xml()}} and
+opened in other applications such as JOSM.
+
+The results are  ready to send the editions to the servers with \code{\link[=osm_diff_upload_changeset]{osm_diff_upload_changeset()}}.
 }
 \description{
 Prepare data to update tags, members and/or latitude and longitude.
