### 2024-05-18 d41371b
Fix tests for errors in server responses
