### 2024-05-18 054a021
Enable tests for fetch_objects_batches. Why it failed without set_osmapi_connection()?    Error in `stop_request(req)`: An unexpected request was made:    GET https://master.apis.dev.openstreetmap.org/api/0.6/relations?relations=...
   Expected mock file: osm.org/api/0.6/relations-13eee2.*

Call to generate mock files is
GET https://api.openstreetmap.org/api/0.6/relations?relations=...
but that was the case in many other tests
