### 2024-05-20 #633 @jonthegeek
Estimated hours spent reviewing: 1 hr (so far)

---

This package looks great!
There's a lot of functionality here to dig into!

A few things to clean up:

These functions are still missing `@examples` in their roxygen blocks: `osm_create_gpx`, `osm_delete_gpx`, `osm_redaction_object`, `osm_update_gpx`.
I know that was discussed previously, but you need to remove the `@examples` roxygen tag if it is empty.
If at all possible, I recommend checking the [Examples section of R Packages (2e)](https://r-pkgs.org/man.html#sec-man-examples) for ideas about how to show *something* for these functions.
Users often scroll down to the examples before reading documentation, so having *something* there to show how the code is intended to work can be helpful, even if it's somewhat contrived (and/or can't actually run).

When I run the tests, there is a lot of "unexpected" output that makes it difficult to see what is happening.
It looks like everything is working in the end, but it would be helpful if you could capture or suppress any expected output, and eliminate anything that shouldn't happen.
For example, `test-user_data.R` is interrupted by `Error in httr2::req_perform(req) : HTTP 404 Not Found.`
You deal with that error in `osm_get_user_details.R` with a `try()` on lines 67-70.
If you use `tryCatch()`, you could replace those lines and suppress the confusing error message, with something like this:

```r
out <- tryCatch(
  osm_details_user(user_id = user_id, format = format),
  error = function(e) empty_user()
)
```

If, on the other hand, you *intend* for the user to see an error in that case, instead I'd wrap the call in `expect_error()`, so it doesn't trigger confusing prints when the test runs, similar to this:

```r
  with_mock_dir("mock_details_user_empty", {
    expect_error({empty_usr <- osm_get_user_details(user_id = 2)})
  })
```

I haven't tracked down the source of the data.frames that print in the test output, but they make it even harder to follow the tests.
I would definitely prefer for those to be suppressed as well!

If you can clean up those pieces, it will make it much easier to see what is supposed to happen as I work through the review!
