### 2024-05-21 #633 @adamhsparks
Thank you, @jonthegeek, for your review and thank you @maurolepore for picking up the slack here. I'd missed the conversation on this somehow in my notifications.

@jonthegeek, we'll wait for @ccamara to submit his review so you'll have time to revise yours or add additional thoughts for a while now.
