### 2024-05-21 5f351da
Reduce output in the test by using expect_shapshot in print method tests
Feedback from @jonthegeek at https://github.com/ropensci/software-review/issues/633#issuecomment-2121192135
