+++ tests/testthat/_snaps/elements.md
@@ -1,0 +1,1115 @@
+# osm_read_object works
+
+    Code
+      print(x)
+    Output
+        type       id visible version changeset           timestamp   user     uid
+      1 node 35308286    TRUE      15  63949815 2018-10-28 13:07:37 ManelG 2491053
+               lat       lon members
+      1 42.5189047 2.4565596    NULL
+                                                                                       tags
+      1 5 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
+
+---
+
+    Code
+      print(x)
+    Output
+        type       id visible version changeset           timestamp     user      uid
+      1  way 13073736    TRUE      21 114656477 2021-12-07 11:27:19 jmaspons 11725140
+         lat  lon                                                         members
+      1 <NA> <NA> 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
+                                                                                       tags
+      1 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
+
+---
+
+    Code
+      print(x)
+    Output
+            type    id visible version changeset           timestamp  user     uid
+      1 relation 40581    TRUE      53 135982205 2023-05-11 15:10:18 l2212 1775455
+         lat  lon                                                         members
+      1 <NA> <NA> 20 members: node/72994199/admin_centre, way/165579897/outer,...
+                                                                                       tags
+      1 22 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
+
+# osm_history_object works
+
+    Code
+      print(x)
+    Output
+         type       id visible version changeset           timestamp     user     uid
+      1  node 35308286    TRUE       1    292412 2007-09-02 14:55:33  Skywave   10927
+      2  node 35308286    TRUE       2     39997 2007-12-02 19:28:54  Skywave   10927
+      3  node 35308286    TRUE       3     39997 2007-12-02 19:29:01  Skywave   10927
+      4  node 35308286    TRUE       4    131335 2007-12-06 21:08:56  Skywave   10927
+      5  node 35308286    TRUE       5    544716 2008-01-05 17:55:28  Skywave   10927
+      6  node 35308286    TRUE       6    433656 2008-08-28 00:14:47  Skywave   10927
+      7  node 35308286    TRUE       7     61302 2008-10-04 20:55:22  Skywave   10927
+      8  node 35308286    TRUE       8    731089 2009-03-03 20:23:52   Eric S   45284
+      9  node 35308286    TRUE       9   4188890 2010-03-21 10:17:52 petrovsk   90394
+      10 node 35308286    TRUE      10   5978159 2010-10-07 13:28:57  Skywave   10927
+      11 node 35308286    TRUE      11   5980618 2010-10-07 19:16:33  Skywave   10927
+      12 node 35308286    TRUE      12   9467521 2011-10-04 08:48:09  WernerP  315015
+      13 node 35308286    TRUE      13  13343795 2012-10-03 07:49:42      FMB  316958
+      14 node 35308286    TRUE      14  18407951 2013-10-17 18:21:18    janes  428521
+      15 node 35308286    TRUE      15  63949815 2018-10-28 13:07:37   ManelG 2491053
+                lat       lon members
+      1  42.5178841 2.4567366    NULL
[…]
