### 2024-05-21 750298a
tryCatch to suppress error message in tests
Feedback from @jonthegeek at https://github.com/ropensci/software-review/issues/633#issuecomment-2121192135
