### 2024-05-21 3f48c02
Messages in a single call per function as expect_message just catch the first
Catch all tests which throw messages with expect_message()
