+++ tests/testthat/test-elements.R
@@ -96,12 +96,15 @@ test_that("edit OSM object works", {
 
 
   with_mock_dir("mock_edit_objects", {
-    changeset_id <- osm_create_changeset(
-      comment = "Test object creation",
-      created_by = "osmapiR", # avoid changes in calls when updating version
-      source = "Imagination",
-      hashtags = "#testing;#osmapiR",
-      verbose = TRUE
+    expect_message(
+      changeset_id <- osm_create_changeset(
+        comment = "Test object creation",
+        created_by = "osmapiR", # avoid changes in calls when updating version
+        source = "Imagination",
+        hashtags = "#testing;#osmapiR",
+        verbose = TRUE
+      ),
+      "New changeset with id = "
     )
 
 
@@ -364,7 +367,7 @@ test_that("osm_fetch_objects works", {
   )
   osm_ids <- as.character(osm_ids)
 
-  set_osmapi_connection() # TODO: why it fails without changing the server?
+  expect_message(set_osmapi_connection(), "Logged out from ") # TODO: why it fails without changing the server?
   # Tests work when running interactively but fail in R CMD check:
   #
   #   Error in `stop_request(req)`: An unexpected request was made:
@@ -399,7 +402,7 @@ test_that("osm_fetch_objects works", {
     expect_identical(id, osm_ids)
     expect_false(any(duplicated(id)))
   })
-  set_osmapi_connection("testing") # TODO
+  expect_message(set_osmapi_connection("testing"), "Logged out from") # TODO
 
   # methods
   lapply(fetch, function(x) expect_snapshot(print(x)))
@@ -500,11 +503,14 @@ test_that("osm_redaction_object works", {
   obj <- osmapi_objects(x, tag_columns = "name")
 
   with_mock_dir("mock_redact_object", {
-    changeset_id <- osm_create_changeset(
-      comment = "Test object redaction",
-      created_by = "osmapiR", # avoid changes in calls when updating version
-      hashtags = "#testing;#osmapiR",
-      verbose = TRUE
+    expect_message(
+      changeset_id <- osm_create_changeset(
+        comment = "Test object redaction",
+        created_by = "osmapiR", # avoid changes in calls when updating version
+        hashtags = "#testing;#osmapiR",
+        verbose = TRUE
+      ),
+      "New changeset with id = "
     )
 
     node_id <- osm_create_object(x = obj, changeset_id = changeset_id)
