+++ tests/testthat/test-map_notes.R
@@ -15,6 +15,7 @@ class_column_comments <- list(
   action = "character", text = "character", html = "character"
 )
 
+
 ## Retrieving notes data by bounding box: `GET /api/0.6/notes` ----
 
 test_that("osm_read_bbox_notes works", {
@@ -89,8 +90,10 @@ test_that("osm_read_note works", {
   lapply(read_note[c("xml", "rss", "gpx")], function(x) expect_true(xml2::xml_length(x) == 1))
   lapply(read_notes[c("xml", "rss", "gpx")], function(x) {
     expect_true(xml2::xml_length(x) == 2)
-    try(expect_identical(xml2::xml_child(x, 1), xml2::xml_child(x, 2)))
-    # TODO: fix added namespaces in 2on node
+    tryCatch(
+      expect_identical(xml2::xml_child(x, 1), xml2::xml_child(x, 2)),
+      error = function(e) message("TODO: fix added namespaces in the 2on node R/osm_get_notes.R/osm_get_notes()")
+    )
   })
 
   # methods
