### 2024-05-21 70e462e
Add vignette to edit gps traces
