+++ pkgdown.yml
@@ -2,8 +2,9 @@ pandoc: 3.1.11
 pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
+  how_to_edit_gps_traces: how_to_edit_gps_traces.html
   how_to_edit_osm: how_to_edit_osm.html
-last_built: 2024-05-18T09:26Z
+last_built: 2024-05-21T16:07Z
 urls:
   reference: https://jmaspons.github.io/osmapiR/reference
   article: https://jmaspons.github.io/osmapiR/articles
