### 2024-05-21 jmaspons/osmapiR#22 @jmaspons
Fix osm_update_gpx()
See https://github.com/openstreetmap/openstreetmap-website/issues/4821

Update https://github.com/jmaspons/osmapiR/blob/main/vignettes/how_to_edit_gps_traces.Rmd
