### 2024-05-21 #633 @jmaspons
Thanks for the feedback, @jonthegeek ! I fixed your comments in https://github.com/jmaspons/osmapiR/pull/21 and found a new issue https://github.com/jmaspons/osmapiR/issues/22
