### 2024-05-23 6446c98
Correct API usage for osm_update_gpx() + test
FIX #22
