+++ tests/testthat/test-gps_traces.R
@@ -58,26 +58,30 @@ test_that("osm_get_points_gps works", {
 test_that("edit gpx works", {
   gpx_path <- test_path("sample_files", "sample.gpx")
 
-  # with_mock_dir("mock_edit_gpx", {
-  #   # TODO: Error in x$value : $ operator is invalid for atomic vectors
-  #   # FIXED with https://github.com/nealrichardson/httptest2/pull/40
-  #   ## Create: `POST /api/0.6/gpx/create` ----
-  #   create_trace <- osm_create_gpx(
-  #     file = gpx_path,
-  #     description = "Test create gpx with osmapiR.",
-  #     tags = c("testing", "osmapiR")
-  #   )
-  #
-  #   ## Update: `PUT /api/0.6/gpx/#id` ----
-  #   # upd_trace <- osm_update_gpx(gpx_id = create_trace, file = gpx_path)
-  #   # TODO: HTTP 400 Bad Request. Cannot parse valid trace from xml string
-  #
-  #   ## Delete: `DELETE /api/0.6/gpx/#id` ----
-  #   del_trace <- osm_delete_gpx(gpx_id = create_trace)
-  # })
-  #
+  with_mock_dir("mock_edit_gpx", {
+    # TODO: Error in x$value : $ operator is invalid for atomic vectors
+    # FIXED with https://github.com/nealrichardson/httptest2/pull/40
+    ## Create: `POST /api/0.6/gpx/create` ----
+    # gpx_id <- osm_create_gpx(
+    #   file = gpx_path,
+    #   description = "Test create gpx with osmapiR.",
+    #   tags = c("testing", "osmapiR")
+    # )
+    gpx_id <- 2283
+
+    ## Update: `PUT /api/0.6/gpx/#id` ----
+    upd_trace <- osm_update_gpx(
+      gpx_id = gpx_id, name = "Upd.gpx", description = "Test update gpx with osmapiR",
+      tags = c("testing", "osmapiR", "updated"), visibility = "identifiable"
+    )
+
+    ## Delete: `DELETE /api/0.6/gpx/#id` ----
+    del_trace <- osm_delete_gpx(gpx_id = gpx_id)
+  })
+
   # expect_type(create_trace, "character")
-  # expect_null(del_trace)
+  expect_s3_class(upd_trace, "data.frame")
+  expect_null(del_trace)
 })
 
 
