### 2024-05-23 e702218
Deploying to gh-pages from @ jmaspons/osmapiR@6446c98fb882aa62c312b76cb9ff6bfc6d2260d9 🚀
