### 2024-05-23 dfdf744
Add id placeholders if missing in osmchange_create()
Make .osc files compatible with JOSM. Otherwise, JOSM will fail to open
even if the file is fine for OSM servers.
