+++ vignettes/how_to_edit_gps_traces.Rmd
@@ -35,22 +35,14 @@ gpx_id <- osm_create_gpx(
 )
 ```
 
-Update the gpx trace with the content of another file:
-
-*TODO: <https://github.com/openstreetmap/openstreetmap-website/issues/4821>*
+Update the gpx trace metadata:
 
 ```r
-gpx_upd_path <- system.file("extdata", "St_Geroni_70nodes.gpx", package = "osmapiR")
-upd_trace <- osm_update_gpx(gpx_id = gpx_id, file = gpx_upd_path)
-#> ! HTTP 500 Internal Server Error.
-#> • The server encountered an internal error or
-#> misconfiguration and was unable to complete
-#> your request.
-#> • Please contact the server administrator at
-#>  webmaster@openstreetmap.org to inform them of the time this error occurred,
-#>  and the actions you performed just before this error.
-#> • More information about this error may be available
-#> in the server error log.
+ori_trace <- osm_get_gpx_metadata(gpx_id = gpx_id)
+upd_trace <- osm_update_gpx(
+  gpx_id = gpx_id, name = "updated.gpx", description = "Test update gpx with osmapiR",
+  tags = c("testing", "osmapiR", "updated"), visibility = "identifiable"
+)
 ```
 
 Delete the gpx trace:
