### 2024-05-23 067640e
Update wiki text to the last version https://wiki.openstreetmap.org/w/index.php?title=API_v0.6&oldid=2711808
