+++ tests/testthat/test-gps_traces.R
@@ -15,10 +15,16 @@ class_columns <- list(
 
 test_that("osm_get_points_gps works", {
   pts_gps <- list()
+  xml_gps <- list()
   with_mock_dir("mock_get_points_gps", {
     pts_gps$private <- osm_get_points_gps(bbox = c(-0.4789191, 38.1662652, -0.4778007, 38.1677898))
     pts_gps$public <- osm_get_points_gps(bbox = c(-0.6430006, 38.1073445, -0.6347179, 38.1112953))
+
+    xml_gps$private <- osm_get_points_gps(bbox = c(-0.4789191, 38.1662652, -0.4778007, 38.1677898), format = "gpx")
+    xml_gps$public <- osm_get_points_gps(bbox = c(-0.6430006, 38.1073445, -0.6347179, 38.1112953), format = "gpx")
   })
+
+  lapply(xml_gps, expect_s3_class, "xml_document")
 
   lapply(pts_gps, expect_type, "list")
   lapply(pts_gps, expect_s3_class, "osmapi_gpx")
@@ -59,15 +65,12 @@ test_that("edit gpx works", {
   gpx_path <- test_path("sample_files", "sample.gpx")
 
   with_mock_dir("mock_edit_gpx", {
-    # TODO: Error in x$value : $ operator is invalid for atomic vectors
-    # FIXED with https://github.com/nealrichardson/httptest2/pull/40
     ## Create: `POST /api/0.6/gpx/create` ----
-    # gpx_id <- osm_create_gpx(
-    #   file = gpx_path,
-    #   description = "Test create gpx with osmapiR.",
-    #   tags = c("testing", "osmapiR")
-    # )
-    gpx_id <- 2283
+    gpx_id <- osm_create_gpx(
+      file = gpx_path,
+      description = "Test create gpx with osmapiR.",
+      tags = c("testing", "osmapiR")
+    )
 
     ## Update: `PUT /api/0.6/gpx/#id` ----
     upd_trace <- osm_update_gpx(
@@ -79,7 +82,8 @@ test_that("edit gpx works", {
     del_trace <- osm_delete_gpx(gpx_id = gpx_id)
   })
 
-  # expect_type(create_trace, "character")
+  expect_type(gpx_id, "character")
+  expect_match(gpx_id, "^[0-9]+$")
   expect_s3_class(upd_trace, "data.frame")
   expect_null(del_trace)
 })
