### 2024-05-24 919672e
Update test-coverage.yaml due to https://github.com/r-lib/actions/issues/834
