+++ .github/workflows/rhub.yaml
@@ -15,7 +15,7 @@ on:
       config:
         description: 'A comma separated list of R-hub platforms to use.'
         type: string
-        default: 'linux,windows,macos'
+        default: 'linux,windows,macos,macos-arm64'
       name:
         description: 'Run name. You can leave this empty now.'
         type: string
