### 2024-05-25 b9597c2
Fix parsing changesets from osm_query_changesets() with empty and non empty changesets
Empty changesets don't have attributes "min_lat", "min_lon", "max_lat", "max_lon"
and give problems when mixed in the do.call(rbind, changeset_attrs)
