### 2024-06-08 62cffbb
Allow more than 100 changesets for osm_query_changesets()
* Fix requests parameter time when time and time_2 parameters are present
* .osm_query_changesets() is an intern function and expects correct input
* set the default limit from getOptions(osmapir.api_capabilities)
