+++ R/osm_query_changesets.R
@@ -1,0 +1,263 @@
+#' Query changesets
+#'
+#' This is an API method for querying changesets. It supports querying by different criteria.
+#'
+#' @param bbox Find changesets within the given bounding box coordinates (`left,bottom,right,top`).
+#' @param user Find changesets by the user with the given user id (numeric) or display name (character).
+#' @param time Find changesets **closed** after this date and time. See details for the valid formats.
+#' @param time_2 find changesets that were **closed** after `time` and **created** before `time_2`. In other words, any
+#'   changesets that were open **at some time** during the given time range `time` to `time_2`.
+#' @param open If `TRUE`, only finds changesets that are still **open** but excludes changesets that are closed or have
+#'   reached the element limit for a changeset (10,000 at the moment `osm_capabilities()$api$changesets`).
+#' @param closed If `TRUE`, only finds changesets that are **closed** or have reached the element limit.
+#' @param changeset_ids Finds changesets with the specified ids.
+#' @param order If `"newest"` (default), sort newest changesets first. If `"oldest"`, reverse order.
+#' @param limit Specifies the maximum number of changesets returned. 100 as the default value.
+#' @param format Format of the output. Can be `R` (default), `xml`, or `json`.
+#' @param tags_in_columns If `FALSE` (default), the tags of the changesets are saved in a single list column `tags`
+#'   containing a `data.frame` for each changeset with the keys and values. If `TRUE`, add a column for each key.
+#'   Ignored if `format != "R"`.
+#'
+#' @details
+#' Where multiple queries are given the result will be those which match all of the requirements. The contents of the
+#' returned document are the changesets and their tags. To get the full set of changes associated with a changeset, use
+#' [osm_download_changeset()] on each changeset ID individually.
+#'
+#' Modification and extension of the basic queries above may be required to support rollback and other uses we find for
+#' changesets.
+#'
+#' This call returns latest changesets matching criteria. The default ordering is newest first, but you can specify
+#' `order = "oldest"` to reverse the sort order (see
+#' [ordered by `created_at`](https://github.com/openstreetmap/openstreetmap-website/blob/f1c6a87aa137c11d0aff5a4b0e563ac2c2a8f82d/app/controllers/api/changesets_controller.rb#L174)
+#' – see the [current state](https://github.com/openstreetmap/openstreetmap-website/blob/master/app/controllers/api/changesets_controller.rb#L174)).
+#' Reverse ordering cannot be combined with `time`.
+#'
+#' Te valid formats for `time` and `time_2` parameters are anything that
+#' [`Time.parse` Ruby function](https://ruby-doc.org/stdlib-2.7.0/libdoc/time/rdoc/Time.html#method-c-parse) will parse.
+#'
+#' @return
+#' If `format = "R"`, returns a data frame with one OSM changeset per row.
+#'
+#' ## `format = "xml"`
+#' Returns a [xml2::xml_document-class] with the following format:
+#' ``` xml
+#' <osm>
[…]
