+++ tests/testthat/_snaps/changesets.md
@@ -137,9 +137,9 @@
       2             1
       3             2
                                                                                        tags
-      1 6 tags: comment=--- | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/...
-      2 6 tags: comment=--- | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/...
-      3 6 tags: comment=--- | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/...
+      1 6 tags: changesets_count=156 | comment=--- | created_by=iD 2.25.2 | host=https:/...
+      2 6 tags: changesets_count=155 | comment=--- | created_by=iD 2.25.2 | host=https:/...
+      3 6 tags: changesets_count=154 | comment=--- | created_by=iD 2.25.2 | host=https:/...
 
 ---
 
@@ -450,106 +450,106 @@
       99              2
       100             1
                                                                                          tags
-      1   6 tags: comment=18945 | created_by=iD 2.25.2 | host=https://www.openstreetmap.or...
-      2   6 tags: comment=18945 | created_by=iD 2.25.2 | host=https://www.openstreetmap.or...
-      3   6 tags: comment=18945 | created_by=iD 2.25.2 | host=https://www.openstreetmap.or...
-      4   6 tags: comment=11 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      5   6 tags: comment=13 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      6   6 tags: comment=13 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      7   6 tags: comment=13 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      8   6 tags: comment=11 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      9   7 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      10  6 tags: comment=13 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      11  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      12  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      13  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      14  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      15  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      16  6 tags: comment=12 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      17  6 tags: comment=48 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      18  6 tags: comment=45 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      19  6 tags: comment=45 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      20  6 tags: comment=46 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      21  6 tags: comment=45 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
-      22  7 tags: comment=45 | created_by=iD 2.25.2 | host=https://www.openstreetmap.org/e...
[…]
