### 2024-06-09 25df7e4
`order = "oldest"` with `time` parameter not allowed
Avoid API error:
! HTTP 400 Bad Request.
• cannot use order=oldest with time

Remove unused code paths
