### 2024-06-09 c929338
Deploying to gh-pages from @ jmaspons/osmapiR@fea458c75daf555cbfb1f74ecaec414a80b8fef8 🚀
