### 2024-06-09 f580aa7
Deploying to gh-pages from @ jmaspons/osmapiR@ed615e82ac74cbd8f78fb8f0fb6327d8ea13098f 🚀
