### 2024-06-11 #633 @ccamara
I'm happy to submit my review. Full disclaimer: this is my first package review ever

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

I want to thank @jmaspons for their excellent job in doing this useful package which (full disclaimer) I am very likely to be using in the future.

The package implements more than 20 functions and has a function matching every [OSM API call](https://wiki.openstreetmap.org/wiki/API_v0.6#API_calls). All the major functions are well documented, with examples, and have dedicated tests. I couldn't find major issues via automated tools, and I can see that the issues flagged by @jonthegeek's review have been addressed\*.

My main concern is on understanding what this package is (and how it differs from other potentially related packages and OSMs APIs, such as osmdata) and what it can do. While I can significant efforts in documentation functions and readme, I must say that I've been an active OSM contributor since 2014, I'm familiar with all the concepts (changesets, history, gpx traces...) and other APIs, such as overpass, but even with that, I struggle to get a full picture of what this package can do, and more importantly, when should be used. IMHO, this is important given that OSM discourages automated edits. It could be argued that the audience of this package should be familiar with OSM API already, and therefore, they know the answers to those questions, and I'd be OK with that. However, my experience is that more often than not, people are not familiar with most of OSM nuissances and would be looking for a package to "work with OSM" searching on a search engine for "OpenStreetMap R package" or "OSM R package" (or similar queries) and may get here. IMHO, a disclaimer, or note for those people either would be appreciated.

That links to my next item: I find the example on the readme file is useful, well explained, relevant and reproducible, but I wonder how representative it is, given that `osmapiR` can do way more than that. Again, I'd suggest having an indication on how to get started, which can be a link to the aforementioned help page.

So, as a wrap up, I would suggest improving the readme file with the following suggestions:

-   Improve the package description, making it slightly more clearer by extending the succinct "Implements OpenStreetMap API for R." to something like "An R interface to OpenStreetMap API v0.6 for fetching and saving raw geodata from/to the OpenStreetMap database" (copied from <https://wiki.openstreetmap.org/wiki/API>) to clarify the scope of the package and to make it even clearer disctinction with other OSM's APIs (i.e. overpass).

-   Make a note on the scope (when to use or not to use it), and adding any considerations when using it, i.e. mention about automated edits, when putting together disperse information in other parts of the documentation such as API usage (modifying user agent, using a test

-   Outline what the package can do, i.e., by mentioning that can retrieve and update OSM objects, changesets, changeset's discussions, notes, GPS traces and users.

-   Indicate how to get started with the package, beyond the example

That said, I acknowledge that A) most of it has been addressed in the `?osmapiR` , so it could either be adapted from there and indicate prominently that this is the place to get started, and B) I do no think this is a stopper and could be done after the package is reviewed and accepted (I would kindly contribute with the documentation).

\* I didn't notice any warnings when running the tests, and I can see empty examples being deleted, but I can see `osm_create_gpx` . `osm_update_gpx`and `osm_delete_gpx do` not include an empty example, pointing to a dedicated vignette instead, which I believe is missing.
