### 2024-06-11 #633 @adamhsparks
> Apologies @jmaspons @ccamara and @adamhsparks for checks not appearing previously. The call was issued exactly when our system was in rebuild mode, as mentioned in our [_Dev Guide_ (4th point in the list)](https://devguide.ropensci.org/softwarereview_editor.html#upon-submission). I'll manually trigger the check results now.

Thanks, @mpadge. I admit, I didn't read that bullet point before. Now I know!
