### 2024-06-11 c7de554
Limits for osm_search_notes() from getOption("osmapir.api_capabilities")$api$notes
TODO: calls in batches by date range as in osm_query_changesets()
