### 2024-06-11 #633 @mpadge
Another update for @ccamara @adamhsparks. Carlos, you were unable to start the package checks because Adam's' original call to the bot was `assign ccamara as reviewer` when it should have been `assign @ccamara as reviewer`. This logged the reviewer as "ccamara", which of course has no meaning in GitHub terms. I've updated the template above to insert the missing `@` symbols, so now you're officially recognised as a reviewer here. Adam, you can add this to your list of ...
> Now I know!

... even more :rofl:
