### 2024-06-11 e3badc1
Fix typo and improve message
