+++ R/osmapi_gps_traces.R
@@ -427,7 +427,8 @@ osm_get_data_gpx <- function(gpx_id, format) {
     if (length(out) > 1) {
       warning(
         "Unexpected output format at osm_get_data_gpx().",
-        "Please, open and issue with with the `gpx_id` at https://github.com/jmaspons/osmapiR/issues"
+        "Please, open and issue with the `gpx_id` or the original file if the gpx is not public ",
+        "at https://github.com/jmaspons/osmapiR/issues"
       )
     } else {
       attrs <- attributes(out)
