### 2024-06-11 #633 @jmaspons
Thanks for the review, @ccamara! I will try to improve the README with your feedback and, as you would like to contribute with the docs, I will ask you to review the PR.

> but I can see `osm_create_gpx` . `osm_update_gpx`and `osm_delete_gpx do` not include an empty example, pointing to a dedicated vignette instead, which I believe is missing.

I will check what's missing to include the vignette in the package, but you can check it at https://jmaspons.github.io/osmapiR/articles/how_to_edit_gps_traces.html
