### 2024-06-11 c2289a2
Install vignettes when using remotes::install_github()
Ctrl +  Shift + B in Rstudio doesn't install vignettes
Feedback from @ccamara at https://github.com/ropensci/software-review/issues/633#issuecomment-2121192135https://github.com/ropensci/software-review/issues/633#issuecomment-2160182185
