### 2024-06-11 #633 @ccamara
@jmaspons a pleasure! Thanks for fixing the vignettes part so quickly (I detected another problem, though: `rmarkdown` is now required to build vignettes, so needs to be specified (see PR https://github.com/jmaspons/osmapiR/pull/24) or added as a required dependency.

Yes, to be honest, when I started reviewing your package yesterday, your name sounded familiar, and I realised that it was a happy coincidence that our paths crossed on OSM a the day before! (For the record: that doesn't mean any conflict of interest in my review).

@mpadge thank you very much for the friendly explanation. You're building such a great community here by treating people so well!

@adamhsparks the experience couldn't have been more positive for me and I'm learning a lot here. I'm not sure if there's anything else I need to do as a reviewer.
