### 2024-06-12 6579e9e
Improve documentation
Better package description
Move examples from README to the Get started Vignette

Feedback from @ccamara at https://github.com/ropensci/software-review/issues/633#issuecomment-2121192135https://github.com/ropensci/software-review/issues/633#issuecomment-2160182185
