+++ README.md
@@ -15,8 +15,13 @@ Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repost
 [![](https://badges.ropensci.org/633_status.svg)](https://github.com/ropensci/software-review/issues/633)
 <!-- badges: end -->
 
-Implements [OpenStreetMap
-API](https://wiki.openstreetmap.org/wiki/API_v0.6) for R.
+An R interface to [OpenStreetMap API
+v0.6](https://wiki.openstreetmap.org/wiki/API_v0.6) for fetching and
+saving raw geodata from/to the OpenStreetMap database. It allows to
+access other OSM data than the maps data (map notes, GPS traces,
+changelogs and users). To access the OSM map data for purposes other
+than editing or exploring the history of the objects see [Related
+packages](README.md#related-packages).
 
 You are responsible for following the [API Usage
 Policy](https://operations.osmfoundation.org/policies/api/). You can
@@ -27,109 +32,29 @@ modify the user agent of the requests by setting the option
 options(osmapir.user_agent = "my new user agent")
 ```
 
+Respect and follow the [standards and
+conventions](https://wiki.openstreetmap.org/wiki/Editing_Standards_and_Conventions)
+of the OpenStreetMap community. If you plan to do automated edits, check
+the [Automated Edits code of
+conduct](https://wiki.openstreetmap.org/wiki/Automated_Edits_code_of_conduct).
+
 ## Installation
 
 You can install the development version of osmapiR from
-[GitHub](https://github.com/) with:
+[GitHub](https://github.com) with:
 
 ``` r
 # install.packages("remotes")
 remotes::install_github("jmaspons/osmapiR", build_vignettes = TRUE)
 ```
 
-## Examples
+## Getting started
 
-This is a basic example which shows how to get map data:
+For an overview of the functionalities, check `?osmapiR-package` or the
+[web](https://jmaspons.github.io/osmapiR/reference/index.html).
 
-``` r
-library(osmapiR)
-#> Data (c) OpenStreetMap contributors, ODbL 1.0. https://www.openstreetmap.org/copyright
-```
-
-``` r
-
-# For testing edition without breaking the OSM data,
-# make calls to `https://master.apis.dev.openstreetmap.org`
-
-# set_osmapi_connection(server = "testing") # lacks data
-
-bbox <- c(2.4166059, 42.5945594, 2.4176574, 42.5962101)
-
-## Download objects by bounding box
-osm_objs <- osm_bbox_objects(bbox = bbox, tags_in_columns = TRUE)
-
-## View history of an object
-sel <- osm_objs$`name:ca` %in% "Abadia de Sant Miquel de Cuixà"
-obj <- osm_objs[sel, ]
-
-obj_history <- osm_history_object(osm_type = obj$osm_type, osm_id = obj$osm_id) # tags in a list column
-obj_history
-#>   type       id visible version changeset           timestamp     user    uid
-#> 1  way 50343004    TRUE       1   3882565 2010-02-15 12:14:11  Skywave  10927
-#> 2  way 50343004    TRUE       2  13314595 2012-09-30 19:52:28 petrovsk  90394
-#> 3  way 50343004    TRUE       3  53623614 2017-11-08 22:08:33    JFK73 662440
-#> 4  way 50343004    TRUE       4 103004865 2021-04-15 15:41:02 petrovsk  90394
-#>    lat  lon                                                         members
-#> 1 <NA> <NA> 44 nodes: 639618609, 639618608, 639618720, 639618717, 639618...
[…]

[…]
