### 2024-06-12 #633 @jmaspons
PR with improvements in the README and docs based on the review of @ccamara at https://github.com/jmaspons/osmapiR/pull/25
