### 2024-06-12 f8c5e02
Improve docs
Character arguments with `value` -> `"value"`
Examples with internal function osm_fetch_objects() & osm_details_users()
format = "osc" for osmachange output and xml as a synonym
