+++ R/osmapi_gps_traces.R
@@ -55,7 +55,7 @@
 #'   box)— in order to retrieve all of the points for a bounding box that contains more than 5,000 points. When this
 #'   parameter is 0 (zero), the command returns the first 5,000 points; when it is 1, the command returns points
 #'   5,001–10,000, etc.
-#' @param format Format of the output. Can be `R` (default) or `gpx`.
+#' @param format Format of the output. Can be `"R"` (default) or `"gpx"`.
 #'
 #' @note In violation of the [GPX standard](https://www.topografix.com/GPX/1/1/#type_trksegType) when downloading public
 #'   GPX traces through the API, all waypoints of non-trackable traces are randomized (or rather sorted by lat/lon) and
@@ -322,7 +322,7 @@ osm_delete_gpx <- function(gpx_id) {
 #' Otherwise only usable by the owner account and requires authentication.
 #'
 #' @param gpx_id The track id represented by a numeric or a character value.
-#' @param format Format of the output. Can be `R` (default) or `xml`.
+#' @param format Format of the output. Can be `"R"` (default) or `"xml"`.
 #'
 #' @return
 #' If `format = "R"`, returns a data frame with one trace per row. If `format = "xml"`, returns a
@@ -381,9 +381,9 @@ osm_get_metadata_gpx <- function(gpx_id, format = c("R", "xml")) {
 #'
 #' @param gpx_id The track id represented by a numeric or a character value.
 #' @param format Format of the output. If missing (default), the response will be the exact file that was uploaded.
-#'   If `R`, a `data.frame`.
-#'   If `gpx`, the response will always be a GPX format file.
-#'   If `xml`, a `XML` file in an undocumented format.
+#'   If `"R"`, a `data.frame`.
+#'   If `"gpx"`, the response will always be a GPX format file.
+#'   If `"xml"`, a `"xml"` file in an undocumented format.
 #'
 #' @note If you request refers to a multi-file archive the response when you force gpx or xml format will consist of a
 #'   non-standard simple concatenation of the files.
@@ -470,7 +470,7 @@ osm_get_data_gpx <- function(gpx_id, format) {
 #'
 #' Use this to get a list of GPX traces owned by the authenticated user. Requires authentication.
 #'
-#' @param format Format of the output. Can be `R` (default) or `xml`.
+#' @param format Format of the output. Can be `"R"` (default) or `"xml"`.
 #'
 #' @return
 #' If `format = "R"`, returns a data frame with one trace per row. If `format = "xml"`, returns a
