### 2024-06-12 bc06a4d
Deploying to gh-pages from @ jmaspons/osmapiR@f8c5e029b67cd3216c46fc1456d53ee3a6032aac 🚀
