### 2024-06-12 53f3b3a
Deal with 0 row input in osmapi_objects() and osmchange_*()
