+++ tests/testthat/test-osmapiR_objects.R
@@ -41,7 +41,9 @@ test_that("osmapi_objects works", {
     lapply(x$tags, function(y) expect_false(any(is.na(y$value))))
   })
 
+
   # keep_na_tags = TRUE
+
   objs_na <- list()
   objs_na$tag_ch <- osmapi_objects(x, tag_columns = c("type.1", "name"), keep_na_tags = TRUE)
   objs_na$tag_num <- osmapi_objects(x, tag_columns = 6:5, keep_na_tags = TRUE)
@@ -58,5 +60,27 @@ test_that("osmapi_objects works", {
       exact = TRUE
     )
     expect_true(any(vapply(x$tags, function(y) any(is.na(y$value)), FUN.VALUE = logical(1))))
+  })
+
+
+  # 0 row input
+
+  x_empty <- x[logical(), ]
+  objs_empty <- list()
+  objs_empty$tag_ch <- osmapi_objects(x_empty, tag_columns = c("type.1", "name"))
+  objs_empty$tag_bool <- osmapi_objects(x_empty, tag_columns = c(FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE))
+  objs_empty$tag_ch_named <- osmapi_objects(x_empty, tag_columns = c(type = "type.1"))
+  objs_empty$mis <- osmapi_objects(x_empty)
+
+  x_empty$tags <- list()
+  objs_empty$tags <- osmapi_objects(x_empty)
+
+  lapply(objs_empty, function(x) {
+    expect_s3_class(
+      validate_osmapi_objects(x, commited = FALSE),
+      class = c("osmapi_objects", "data.frame"),
+      exact = TRUE
+    )
+    expect_identical(nrow(x), 0L)
   })
 })
