### 2024-06-12 f411825
Deploying to gh-pages from @ jmaspons/osmapiR@5436d5978d9b056968c4dbbcd57a1d30e011bffa 🚀
