### 2024-06-12 #633 @jmaspons
> Hey all, I (finally) read an email from @adamhsparks about this. Just to say: would be happy to review. Looks really interesting and can provide feedback if that would be useful.

Should we wait for the review of @Robinlovelace, @adamhsparks  ? He offered to review the package but has not been assigned as a reviewer. I'm not in a hurry, and I'm happy with the feedback. As soon as you give me the OK, I will submit the package to CRAN.
