### 2024-06-13 #633 @jmaspons
Ok @adamhsparks. From my side, if there is no new feedback from reviewers, I would call it done with https://github.com/jmaspons/osmapiR/pull/21 and https://github.com/jmaspons/osmapiR/pull/25
