### 2024-06-13 ed056f2
Vectorized version of osm_close_note() & osm_reopen_note()
