+++ man/osm_close_note.Rd
@@ -1,28 +1,28 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/osmapi_map_notes.R
+% Please edit documentation in R/osm_status_notes.R
 \name{osm_close_note}
 \alias{osm_close_note}
 \alias{osm_reopen_note}
-\title{Close or reopen a note}
+\title{Close or reopen notes}
 \usage{
 osm_close_note(note_id)
 
 osm_reopen_note(note_id)
 }
 \arguments{
-\item{note_id}{Note id represented by a numeric or a character value.}
+\item{note_id}{Note ids represented by a numeric or character vector.}
 }
 \value{
-Returns a data frame with the closed map note (same format as \code{\link[=osm_get_notes]{osm_get_notes()}} with \code{format = "R"}).
+Returns a data frame with the closed map notes (same format as \code{\link[=osm_get_notes]{osm_get_notes()}} with \code{format = "R"}).
 }
 \description{
 Requires authentication.
 }
 \section{Functions}{
 \itemize{
-\item \code{osm_close_note()}: Close a note as fixed.
+\item \code{osm_close_note()}: Close notes as fixed.
 
-\item \code{osm_reopen_note()}: Reopen a closed note.
+\item \code{osm_reopen_note()}: Reopen closed notes.
 
 }}
 \examples{
@@ -33,6 +33,7 @@ closed_note <- osm_close_note(note$id)
 closed_note
 reopened_note <- osm_reopen_note(note$id)
 reopened_note
+closed_note <- osm_close_note(note$id) # leave it closed
 }
 }
 \seealso{
