### 2024-06-13 dbb3a3f
Vectorized version of osm_delete_gpx()
