+++ man/osm_delete_gpx.Rd
@@ -1,19 +1,19 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/osmapi_gps_traces.R
+% Please edit documentation in R/osm_delete_gpx.R
 \name{osm_delete_gpx}
 \alias{osm_delete_gpx}
-\title{Delete GPS trace}
+\title{Delete GPS traces}
 \usage{
 osm_delete_gpx(gpx_id)
 }
 \arguments{
-\item{gpx_id}{The track id represented by a numeric or a character value.}
+\item{gpx_id}{The track ids represented by a numeric or a character vector.}
 }
 \value{
 Returns \code{NULL} invisibly.
 }
 \description{
-Use this to delete a GPX file. Only usable by the owner account. Requires authentication.
+Use this to delete GPX files. Only usable by the owner account. Requires authentication.
 }
 \examples{
 vignette("how_to_edit_gps_traces", package = "osmapiR")
