+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   how_to_edit_gps_traces: how_to_edit_gps_traces.html
   how_to_edit_osm: how_to_edit_osm.html
   osmapiR: osmapiR.html
-last_built: 2024-06-13T08:38Z
+last_built: 2024-06-13T08:48Z
 urls:
   reference: https://jmaspons.github.io/osmapiR/reference
   article: https://jmaspons.github.io/osmapiR/articles
