### 2024-06-13 8401ef1
Vectorized version of osm_delete_gpx()
