+++ man/osm_delete_note.Rd
@@ -1,21 +1,21 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/osmapi_map_notes.R
+% Please edit documentation in R/osm_delete_note.R
 \name{osm_delete_note}
 \alias{osm_delete_note}
-\title{Delete a note}
+\title{Delete notes}
 \usage{
 osm_delete_note(note_id, text)
 }
 \arguments{
-\item{note_id}{Note id represented by a numeric or a character value.}
+\item{note_id}{Note ids represented by a numeric or a character vector.}
 
 \item{text}{A non-mandatory comment as text.}
 }
 \value{
-Returns a data frame with the hided map note (same format as \code{\link[=osm_get_notes]{osm_get_notes()}} with \code{format = "R"}).
+Returns a data frame with the hided map notes (same format as \code{\link[=osm_get_notes]{osm_get_notes()}} with \code{format = "R"}).
 }
 \description{
-Hide (delete) a note. This request needs to be done as an authenticated user with moderator role.
+Hide (delete) notes. This request needs to be done as an authenticated user with moderator role.
 }
 \details{
 Use \code{\link[=osm_reopen_note]{osm_reopen_note()}} to make the note visible again.
