### 2024-06-13 1e67bb5
mapply() -> .mapply() microoptimization
Cases in osm_get_objects() require osm type names in list so not word it to change
