### 2024-06-13 #633 @jmaspons
Thanks to everybody involved in the review! There is no NEWS file yet as there is no release. I will submit the package after the review as I expect that the planned changes can fit the current API of the package.

### Response to @jonthegeek

I implemented all the suggestions with https://github.com/jmaspons/osmapiR/pull/21

[x] Reduce output in the tests using expect_snapshot() and the suggested changes to catch expected errors.
[x] Add examples for: `osm_create_gpx`, `osm_delete_gpx` & `osm_update_gpx` in a vignette, and for `osm_redaction_object`.

### Response to @ccamara

I implemented the suggestions with https://github.com/jmaspons/osmapiR/pull/25

[x] Improve the package description in the README.
[x] Add option to install vignettes when using `remotes::github_install()`.
