### 2024-06-13 f3329fb
Download all pages osm_get_points_gps(..., page_number = -1)
The API call make batches of 5.000 points.
