+++ tests/testthat/test-gps_traces.R
@@ -19,9 +19,13 @@ test_that("osm_get_points_gps works", {
   with_mock_dir("mock_get_points_gps", {
     pts_gps$private <- osm_get_points_gps(bbox = c(-0.4789191, 38.1662652, -0.4778007, 38.1677898))
     pts_gps$public <- osm_get_points_gps(bbox = c(-0.6430006, 38.1073445, -0.6347179, 38.1112953))
+    pts_gps$all_pages <- osm_get_points_gps(bbox = "-0.6683636,38.0610674,-0.6388378,38.1", page_number = -1)
 
     xml_gps$private <- osm_get_points_gps(bbox = c(-0.4789191, 38.1662652, -0.4778007, 38.1677898), format = "gpx")
     xml_gps$public <- osm_get_points_gps(bbox = c(-0.6430006, 38.1073445, -0.6347179, 38.1112953), format = "gpx")
+    xml_gps$all_pages <- osm_get_points_gps(
+      bbox = "-0.6683636,38.0610674,-0.6388378,38.1", page_number = -1, format = "gpx"
+    )
   })
 
   lapply(xml_gps, expect_s3_class, "xml_document")
@@ -38,6 +42,10 @@ test_that("osm_get_points_gps works", {
   # Check that time is extracted, otherwise it's 00:00:00 in local time
   lapply(pts_gps$public, function(x) expect_false(unique(strftime(as.POSIXct(x$time), format = "%M:%S") == "00:00")))
 
+  mapply(function(d, x) {
+    # expect_identical(length(d), length(xml2::xml_children(x)))
+    c(length(d), length(xml2::xml_children(x)))
+  }, d = pts_gps, x = xml_gps)
   # methods
   summary_gpx <- lapply(pts_gps, summary)
   lapply(summary_gpx, expect_s3_class, "data.frame")
@@ -45,17 +53,25 @@ test_that("osm_get_points_gps works", {
 
   ## Empty results
 
+  empty_pts <- list()
+  empty_xml <- list()
   with_mock_dir("mock_get_points_gps_empty", {
-    empty_pts_gps <- osm_get_points_gps(bbox = c(-105, -7, -104.9, -6.9))
+    empty_pts$gps <- osm_get_points_gps(bbox = c(-105, -7, -104.9, -6.9))
+    empty_pts$all_pages <- osm_get_points_gps(bbox = c(-105, -7, -104.9, -6.9), page_number = -1)
+
+    empty_xml$gps <- osm_get_points_gps(bbox = c(-105, -7, -104.9, -6.9), format = "gpx")
+    empty_xml$all_pages <- osm_get_points_gps(bbox = c(-105, -7, -104.9, -6.9), page_number = -1, format = "gpx")
   })
 
-  expect_type(empty_pts_gps, "list")
-  expect_s3_class(empty_pts_gps, "osmapi_gpx")
-  expect_length(empty_pts_gps, 0)
+  lapply(empty_pts, expect_type, type = "list")
+  lapply(empty_pts, expect_s3_class, class = "osmapi_gpx")
+  lapply(empty_pts, expect_length, n = 0)
+
+  lapply(empty_xml, expect_s3_class, "xml_document")
 
   # methods
-  summary_gpx <- summary(empty_pts_gps)
-  expect_s3_class(summary_gpx, "data.frame")
+  summary_gpx <- lapply(empty_pts, summary)
+  lapply(summary_gpx, expect_s3_class, class = "data.frame")
 })
 
 
