+++ man/osm_get_points_gps.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/osmapi_gps_traces.R
+% Please edit documentation in R/osm_get_points_gps.R
 \name{osm_get_points_gps}
 \alias{osm_get_points_gps}
 \title{Get GPS Points}
@@ -11,16 +11,16 @@ osm_get_points_gps(bbox, page_number = 0, format = c("R", "gpx"))
 degrees, expressing a valid bounding box. The maximal width (\code{right - left}) and height (\code{top - bottom}) of the
 bounding box is 0.25 degree.}
 
-\item{page_number}{Specifies which group of 5,000 points, or page, to return. Since the command does not return more
-than 5,000 points at a time, this parameter must be incremented —and the command sent again (using the same bounding
-box)— in order to retrieve all of the points for a bounding box that contains more than 5,000 points. When this
-parameter is 0 (zero), the command returns the first 5,000 points; when it is 1, the command returns points
-5,001–10,000, etc.}
+\item{page_number}{Specifies which groups of 5,000 points, or page, to return. The API call does not return more
+than 5,000 points at a time. In order to retrieve all of the points for a bounding box, set \code{page_number = -1}.
+When this parameter is 0 (zero), the command returns the first 5,000 points; when it is 1, the command returns points
+5,001–10,000, etc. A vector is also valid (e.g. \code{0:2} to get the first 3 pages).}
 
 \item{format}{Format of the output. Can be \code{"R"} (default) or \code{"gpx"}.}
 }
 \value{
-If \code{format = "R"}, returns a list of data frames with the points for each trace.
+If \code{format = "R"}, returns a list of data frames with the points for each trace. For public traces, the data frame
+include the attributes \code{name}, \code{desc} and \code{url}
 \subsection{\code{format = "gpx"}}{
 
 Returns a \link[xml2:oldclass]{xml2::xml_document} with the following format:
@@ -62,6 +62,9 @@ pts_gps <- osm_get_points_gps(bbox = c(-0.3667545, 40.2153246, -0.3354263, 40.23
 ## bbox as a character value also works. Equivalent call:
 # pts_gps <- osm_get_points_gps(bbox = "-0.3667545,40.2153246,-0.3354263,40.2364915")
 pts_gps
+
+## get attributes
+lapply(pts_gps, function(x) attributes(x)[c("name", "desc", "url")])
 }
 \seealso{
 Other get GPS' functions: 
