### 2024-06-13 1fc7ba8
Fix concatenation of xml_nodeset for calls in batches
Only the first node of each batch after the first were included.
