+++ R/osm_get_notes.R
@@ -92,8 +92,10 @@ osm_get_notes <- function(note_id, format = c("R", "xml", "rss", "json", "gpx"))
       out <- do.call(rbind, outL)
     } else if (format %in% c("xml", "rss", "gpx")) {
       out <- xml2::xml_new_root(outL[[1]])
-      for (i in seq_len(length(outL) - 1)) {
-        xml2::xml_add_child(out, xml2::xml_child(outL[[i + 1]]))
+      for (i in seq_along(outL[-1]) + 1) {
+        lapply(xml2::xml_children(outL[[i]]), function(node) {
+          xml2::xml_add_child(out, node)
+        })
       }
       # TODO: remove namespaces for format %in% c("rss", "gpx"). xml2::xml_structure(out) [<wpt [lon, lat]> VS <wpt [lon, lat, xmlns]>]
       # xml namespace https://community.rstudio.com/t/adding-nodes-in-xml2-how-to-avoid-duplicate-default-namespaces/84870/
