### 2024-06-19 30784e2
Set core.autocrlf on windows GHA
