### 2024-06-19 29d830d
Unite traces in consecutive pages with the same URL
Private traces without url can be separated in different items in the
result if they get split due to server pagination
