### 2024-06-19 5de12e3
Deploying to gh-pages from @ jmaspons/osmapiR@24c573a4895ed76460de244b60ed6de681300c21 🚀
