### 2024-06-20 85dd0ac
Add all api_capabilities to osmapir.api_capabilities option & use them more
