### 2024-06-21 cae90ee
Fix osmapir.api_capabilities data types (numeric / integer) for meaningful use
