### 2024-06-21 2ba4b2b
Reduce size of the mock files by changing options "default_query_limit", "maximum_query_limit"
Refresh mock files after changes in code
