+++ tests/testthat/_snaps/changesets.md
@@ -146,900 +146,260 @@
     Code
       print(x)
     Output
-                 id          created_at           closed_at  open               user
-      1   137626978 2023-06-22 02:10:24 2023-06-22 02:10:24 FALSE Mementomoristultus
-      2   137626972 2023-06-22 02:09:43 2023-06-22 02:09:44 FALSE Mementomoristultus
-      3   137626970 2023-06-22 02:09:28 2023-06-22 02:09:28 FALSE Mementomoristultus
-      4   137626862 2023-06-22 02:01:25 2023-06-22 02:01:26 FALSE Mementomoristultus
-      5   137626853 2023-06-22 02:00:48 2023-06-22 02:00:48 FALSE Mementomoristultus
-      6   137626849 2023-06-22 02:00:24 2023-06-22 02:00:24 FALSE Mementomoristultus
-      7   137626836 2023-06-22 01:59:16 2023-06-22 01:59:16 FALSE Mementomoristultus
-      8   137626833 2023-06-22 01:58:51 2023-06-22 01:58:51 FALSE Mementomoristultus
-      9   137626827 2023-06-22 01:58:12 2023-06-22 01:58:12 FALSE Mementomoristultus
-      10  137626819 2023-06-22 01:57:49 2023-06-22 01:57:49 FALSE Mementomoristultus
-      11  137626813 2023-06-22 01:57:20 2023-06-22 01:57:20 FALSE Mementomoristultus
-      12  137626807 2023-06-22 01:56:41 2023-06-22 01:56:42 FALSE Mementomoristultus
-      13  137626802 2023-06-22 01:56:13 2023-06-22 01:56:14 FALSE Mementomoristultus
-      14  137626794 2023-06-22 01:55:22 2023-06-22 01:55:22 FALSE Mementomoristultus
-      15  137626788 2023-06-22 01:54:57 2023-06-22 01:54:57 FALSE Mementomoristultus
-      16  137626782 2023-06-22 01:54:22 2023-06-22 01:54:22 FALSE Mementomoristultus
-      17  137626769 2023-06-22 01:53:38 2023-06-22 01:53:38 FALSE Mementomoristultus
-      18  137626762 2023-06-22 01:53:12 2023-06-22 01:53:13 FALSE Mementomoristultus
-      19  137626755 2023-06-22 01:52:35 2023-06-22 01:52:36 FALSE Mementomoristultus
-      20  137626750 2023-06-22 01:52:18 2023-06-22 01:52:18 FALSE Mementomoristultus
-      21  137626743 2023-06-22 01:51:52 2023-06-22 01:51:52 FALSE Mementomoristultus
-      22  137626736 2023-06-22 01:51:21 2023-06-22 01:51:22 FALSE Mementomoristultus
-      23  137626718 2023-06-22 01:50:19 2023-06-22 01:50:19 FALSE Mementomoristultus
-      24  137626713 2023-06-22 01:50:00 2023-06-22 01:50:00 FALSE Mementomoristultus
-      25  137626710 2023-06-22 01:49:40 2023-06-22 01:49:40 FALSE Mementomoristultus
-      26  137626704 2023-06-22 01:49:20 2023-06-22 01:49:20 FALSE Mementomoristultus
-      27  137626693 2023-06-22 01:48:30 2023-06-22 01:48:31 FALSE Mementomoristultus
-      28  137626667 2023-06-22 01:46:44 2023-06-22 01:46:45 FALSE Mementomoristultus
-      29  137626664 2023-06-22 01:46:23 2023-06-22 01:46:23 FALSE Mementomoristultus
[…]
