### 2024-06-21 #633 @ccamara
Hello, I thought I had done it here: https://github.com/ropensci/software-review/issues/633#issuecomment-2163019078 , but I may be missing something. How/Where should I approve any new/same changes? (sorry for the newbie question)
