### 2024-06-21 #633 @jonthegeek
> Hello, I thought I had done it here: [#633 (comment)](https://github.com/ropensci/software-review/issues/633#issuecomment-2163019078) , but I may be missing something. How/Where should I approve any new/same changes? (sorry for the newbie question).
>
> Anyway, and just to clarify: I'm happy with the changes being made and the current status of the project.

There's a [template in the guide](https://devguide.ropensci.org/approval2template.html). I had totally missed that, too. Thanks for the poke, @adamhsparks !
