### 2024-06-21 7780751
Increase version and usethis::use_tidy_description()
