+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: osmapiR
 Title: OpenStreetMap API
-Version: 0.0.0.39
+Version: 0.0.0.40
 Authors@R: 
     person("Joan", "Maspons", , "joanmaspons@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2286-8727"))
@@ -9,7 +9,8 @@ Maintainer: Joan Maspons <joanmaspons@gmail.com>
 Description: Implements OpenStreetMap API v0.6
     (<https://wiki.openstreetmap.org/wiki/API_v0.6>).
 License: GPL (>= 3)
-URL: https://jmaspons.github.io/osmapiR https://github.com/jmaspons/osmapiR
+URL: https://jmaspons.github.io/osmapiR
+    https://github.com/jmaspons/osmapiR
 BugReports: https://github.com/jmaspons/osmapiR/issues
 Imports: 
     curl,
@@ -21,9 +22,11 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0)
+VignetteBuilder: 
+    knitr
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
-X-schema.org-keywords: open street map, openstreetmap, OSM, openstreetmap-api, osmapi, API
-VignetteBuilder: knitr
+X-schema.org-keywords: open street map, openstreetmap, OSM,
+    openstreetmap-api, osmapi, API
