### 2024-06-21 2d0634a
Improve package docs (from README)
