+++ DESCRIPTION
@@ -6,7 +6,8 @@ Authors@R:
     person("Joan", "Maspons", , "joanmaspons@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2286-8727"))
 Maintainer: Joan Maspons <joanmaspons@gmail.com>
-Description: Implements OpenStreetMap API v0.6
+Description: Interface to OpenStreetMap API for fetching and saving data
+    from/to the OpenStreetMap database
     (<https://wiki.openstreetmap.org/wiki/API_v0.6>).
 License: GPL (>= 3)
 URL: https://jmaspons.github.io/osmapiR
