### 2024-06-21 fc41b67
urlchecker::url_check()
Some errors and warnings for links that require authentication or
are recommended by codecov (badge)
