+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 0.0.0.40
 Authors@R: c(
     person("Joan", "Maspons", , "joanmaspons@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2286-8727")),
-    person("Jon", "Harmon", role = "rev"),
+    person("Jon", "Harmon", role = "rev", comment = c(ORCID = "0000-0003-4781-4346")),
     person("Carlos", "Cámara", role = "rev")
   )
 Maintainer: Joan Maspons <joanmaspons@gmail.com>
