### 2024-06-21 d92becf
Deploying to gh-pages from @ jmaspons/osmapiR@9b3970aa7c959fed9c5848841ccfc0a83fb862e7 🚀
