### 2024-06-21 ddd95e2
Fixes in docs found by devtools::check_win_devel()
