+++ DESCRIPTION
@@ -5,8 +5,10 @@ Version: 0.0.0.40
 Authors@R: c(
     person("Joan", "Maspons", , "joanmaspons@gmail.com", role = c("aut", "cre", "cph"),
            comment = c(ORCID = "0000-0003-2286-8727")),
-    person("Jon", "Harmon", role = "rev", comment = c(ORCID = "0000-0003-4781-4346")),
-    person("Carlos", "Cámara", role = "rev")
+    person("Jon", "Harmon", role = "rev",
+           comment = c("Jon reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/633", ORCID = "0000-0003-4781-4346")),
+    person("Carlos", "Cámara", role = "rev",
+           comment = "Carles reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/633")
   )
 Maintainer: Joan Maspons <joanmaspons@gmail.com>
 Description: Interface to OpenStreetMap API for fetching and saving data
