### 2024-06-27 f190a9b
package names, software names and API names in single quotes in title and description
Comment from CRAN review by Benjamin Altmann
