### 2024-06-27 1208bf8
Remove commented code lines in examples
From Benjamin Altmann' comments
