+++ R/osmapi_miscellaneous.R
@@ -244,8 +244,7 @@ osm_capabilities <- function() {
 #'
 #' @examples
 #' map_data <- osm_bbox_objects(bbox = c(1.8366775, 41.8336843, 1.8379971, 41.8344537))
-#' ## bbox as a character value also works. Equivalent call:
-#' # map_data <- osm_bbox_objects(bbox = "1.8366775,41.8336843,1.8379971,41.8344537")
+#' ## bbox as a character value also works (bbox = "1.8366775,41.8336843,1.8379971,41.8344537").
 #' map_data
 osm_bbox_objects <- function(bbox, format = c("R", "xml", "json"), tags_in_columns = FALSE) {
   format <- match.arg(format)
