+++ man/osm_feed_notes.Rd
@@ -21,8 +21,7 @@ RSS Feed of notes in a bbox
 }
 \examples{
 feed_notes <- osm_feed_notes(bbox = c(0.8205414, 40.6686604, 0.8857727, 40.7493377))
-## bbox as a character value also works. Equivalent call:
-# feed_notes <- osm_feed_notes(bbox = "0.8205414,40.6686604,0.8857727,40.7493377")
+## bbox as a character value also works (bbox = "0.8205414,40.6686604,0.8857727,40.7493377").
 feed_notes
 }
 \seealso{
