### 2024-06-27 890f6b6
Restore modified options in examples
