+++ R/osmapiR_connection.R
@@ -30,7 +30,24 @@
 #' @export
 #'
 #' @examples
+#' ori <- get_osmapi_url()
 #' set_osmapi_connection(server = "testing")
+#' get_osmapi_url()
+#' set_osmapi_connection(server = "openstreetmap.org")
+#' get_osmapi_url()
+#'
+#' ## Restore options
+#' if (ori == "https://api.openstreetmap.org") {
+#'   set_osmapi_connection(server = "openstreetmap.org")
+#' } else if (ori == "https://master.apis.dev.openstreetmap.org") {
+#'   set_osmapi_connection(server = "testing")
+#' } else {
+#'   warning(
+#'     "A non standard osmapiR connection detected (", ori,
+#'     "). If you configured manually options like \"osmapir.base_api_url\" or \"osmapir.oauth_id\", ",
+#'     "configure it again."
+#'   )
+#' }
 set_osmapi_connection <- function(server = c("openstreetmap.org", "testing"), cache_authentication) {
   server <- match.arg(server)
 
