+++ DESCRIPTION
@@ -8,7 +8,7 @@ Authors@R: c(
     person("Jon", "Harmon", role = "rev",
            comment = c("Jon reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/633", ORCID = "0000-0003-4781-4346")),
     person("Carlos", "Cámara", role = "rev",
-           comment = "Carles reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/633")
+           comment = c("Carles reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/633", ORCID = "0000-0002-9378-0549"))
   )
 Maintainer: Joan Maspons <joanmaspons@gmail.com>
 Description: Interface to OpenStreetMap API for fetching and saving data
