### 2024-06-27 98b81f2
Refresh autogenerated osmapiR-package docs
