### 2024-06-27 845a778
First CRAN release: 0.1.0
