+++ README.Rmd
@@ -21,6 +21,7 @@ knitr::opts_chunk$set(
 [![pkgdown](https://github.com/jmaspons/osmapiR/actions/workflows/pkgdown.yaml/badge.svg)](https://github.com/jmaspons/osmapiR/actions/workflows/pkgdown.yaml)
 [![Project Status: Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/osmapiR)](https://cran.r-project.org/package=osmapiR)
+[![CRAN Downloads](https://cranlogs.r-pkg.org/badges/grand-total/osmapiR)](https://cran.r-project.org/package=osmapiR)
 [![](https://badges.ropensci.org/633_status.svg)](https://github.com/ropensci/software-review/issues/633)
 <!-- badges: end -->
 
@@ -42,6 +43,12 @@ community. If you plan to do automated edits, check the
 
 
 ## Installation
+
+To install latest CRAN version:
+
+``` r
+install.packages("osmapiR")
+```
 
 You can install the development version of osmapiR from [GitHub](https://github.com) with:
 
