+++ NEWS.md
@@ -1,0 +1,7 @@
+# osmapiR 0.1.0
+
+* Initial CRAN submission implementing calls to all the API endpoints.
+* Server responses are returned as R objects, xml_documents or json lists.
+* Authentication when needed with OAuth2.
+* Pagination in server responses handled internally (#20, #23 & #29).
+* Vectorization of atomic API calls (#18).
