+++ cran-comments.md
@@ -1,0 +1,14 @@
+## R CMD check results
+
+0 errors | 0 warnings | 0 note
+
+* This is a new release.
+
+urlchecker::url_check() errors and warnings for links that require authentication
+
+## Resubmission
+
+Addressed comments by Benjamin Altmann:
+
+* DESCRIPTION: package names, software names and API names in single quotes in title and description
+* Remove commented code lines in examples
