### 2024-06-30 41d0e5d
Deploying to gh-pages from @ jmaspons/osmapiR@1ad47472b852e88bd99819f2f83da177aaa7f593 🚀
