### 2024-06-30 a6d7cee
Deploying to gh-pages from @ jmaspons/osmapiR@6d8f3bc8d5feaf9e857c974b982adf1912064d0b 🚀
