### 2024-07-01 a00c770
Improve tests and fixes
TODO: osm_full_object() lost last object in format = R (test-elements.R:611)
