+++ tests/testthat/mock_search_notes_empty/osm.org/api/0.6/notes/search.rss-0acea5.R
@@ -1,0 +1,54 @@
+structure(list(method = "GET", url = "osm.org/api/0.6/notes/search.rss?q=Visca%20la%20terra%21&display_name=jmaspons&sort=updated_at&order=newest&limit=100", 
+    status_code = 200L, headers = structure(list(`cache-control` = "max-age=0, private, must-revalidate", 
+        vary = "Accept-Language,Origin,Accept-Encoding", `referrer-policy` = "strict-origin-when-cross-origin", 
+        `x-permitted-cross-domain-policies` = "none", `x-xss-protection` = "0", 
+        `content-language` = "en", `x-request-id` = "REDACTED", 
+        `x-frame-options` = "SAMEORIGIN", `x-runtime` = "0.030965", 
+        `x-content-type-options` = "nosniff", `content-security-policy` = "default-src 'self'; child-src 'self'; connect-src 'self' matomo.openstreetmap.org; font-src 'none'; form-action 'self'; frame-ancestors 'self'; frame-src 'self'; img-src 'self' data: www.gravatar.com *.wp.com tile.openstreetmap.org gps.tile.openstreetmap.org *.tile.thunderforest.com tile.tracestrack.com *.openstreetmap.fr matomo.openstreetmap.org https://openstreetmap-user-avatars.s3.dualstack.eu-west-1.amazonaws.com https://openstreetmap-gps-images.s3.dualstack.eu-west-1.amazonaws.com; manifest-src 'self'; media-src 'none'; object-src 'self'; script-src 'self' matomo.openstreetmap.org; style-src 'self' 'nonce-wetny3Icb+43XxppqP1Oh6osbCNplaXB'; worker-src 'none'", 
+        date = "REDACTED", `x-powered-by` = "Phusion Passenger(R) 6.0.22", 
+        `strict-transport-security` = "max-age=31536000; includeSubDomains; preload", 
+        etag = "W/\"98ced57a144a5d64b8ecf359a9bad5bf-br\"", status = "200 OK", 
+        `content-encoding` = "br", `content-length` = "251", 
+        `content-type` = "application/rss+xml; charset=utf-8", 
+        server = "Apache/2.4.54 (Ubuntu)"), class = "httr2_headers"), 
+    body = as.raw(c(0x3c, 0x3f, 0x78, 0x6d, 0x6c, 0x20, 0x76, 
+    0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x3d, 0x22, 0x31, 0x2e, 
+    0x30, 0x22, 0x20, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 
[…]
