+++ tests/testthat/test-user_data.R
@@ -15,9 +15,13 @@ class_columns <- list(
 
 test_that("osm_details_user works", {
   usr <- list()
+  xml_usr <- list()
   with_mock_dir("mock_details_user", {
     usr$usr <- osm_get_user_details(user_id = "11725140")
     usr$mod <- osm_get_user_details(user_id = 61942)
+
+    xml_usr$usr <- osm_get_user_details(user_id = "11725140", format = "xml")
+    xml_usr$mod <- osm_get_user_details(user_id = 61942, format = "xml")
   })
 
   lapply(usr, expect_s3_class, "data.frame")
@@ -30,22 +34,39 @@ test_that("osm_details_user works", {
   # Check that time is extracted, otherwise it's 00:00:00 in local time
   lapply(usr, function(x) expect_false(unique(strftime(as.POSIXct(x$account_created), format = "%M:%S") == "00:00")))
 
+  lapply(xml_usr, expect_s3_class, "xml_document")
+
+  # Compare xml & R
+  mapply(function(d, x) {
+    expect_identical(nrow(d), xml2::xml_length(x))
+  }, d = usr, x = xml_usr)
+
 
   ## Empty results
 
   with_mock_dir("mock_details_user_empty", {
     empty_usr <- osm_get_user_details(user_id = 2)
+    xml_empty_usr <- osm_get_user_details(user_id = 2, format = "xml")
+    json_empty_usr <- osm_get_user_details(user_id = 2, format = "json")
   })
 
   expect_s3_class(empty_usr, "data.frame")
   expect_named(empty_usr, column_users)
-  expect_identical(nrow(empty_usr), 0L)
+
 
   mapply(
     function(x, cl) expect_true(inherits(x, cl)),
     x = empty_usr,
     cl = class_columns[names(empty_usr)]
   )
+
+  expect_s3_class(xml_empty_usr, "xml_document")
+  expect_type(json_empty_usr, "list")
+
+  # Compare xml, json & R
+  expect_identical(nrow(empty_usr), 0L)
+  expect_identical(xml2::xml_length(xml_empty_usr), 0L)
+  expect_identical(length(json_empty_usr$users), 0L)
 })
 
 
