+++ tests/testthat/mock_search_notes_empty/osm.org/api/0.6/notes/search.gpx-0acea5.R
@@ -1,0 +1,42 @@
+structure(list(method = "GET", url = "osm.org/api/0.6/notes/search.gpx?q=Visca%20la%20terra%21&display_name=jmaspons&sort=updated_at&order=newest&limit=100", 
+    status_code = 200L, headers = structure(list(`cache-control` = "max-age=0, private, must-revalidate", 
+        vary = "Accept-Language,Origin", `referrer-policy` = "strict-origin-when-cross-origin", 
+        `x-permitted-cross-domain-policies` = "none", `x-xss-protection` = "0", 
+        `content-language` = "en", `x-request-id` = "REDACTED", 
+        `x-frame-options` = "SAMEORIGIN", `x-runtime` = "0.027490", 
+        `x-content-type-options` = "nosniff", `content-security-policy` = "default-src 'self'; child-src 'self'; connect-src 'self' matomo.openstreetmap.org; font-src 'none'; form-action 'self'; frame-ancestors 'self'; frame-src 'self'; img-src 'self' data: www.gravatar.com *.wp.com tile.openstreetmap.org gps.tile.openstreetmap.org *.tile.thunderforest.com tile.tracestrack.com *.openstreetmap.fr matomo.openstreetmap.org https://openstreetmap-user-avatars.s3.dualstack.eu-west-1.amazonaws.com https://openstreetmap-gps-images.s3.dualstack.eu-west-1.amazonaws.com; manifest-src 'self'; media-src 'none'; object-src 'self'; script-src 'self' matomo.openstreetmap.org; style-src 'self' 'nonce-qjfZxSMy/mJdVDnQtWelOXz31W62al40'; worker-src 'none'", 
+        date = "REDACTED", `x-powered-by` = "Phusion Passenger(R) 6.0.22", 
+        `strict-transport-security` = "max-age=31536000; includeSubDomains; preload", 
+        etag = "W/\"e7f53697295114ce679fe91238e0ec28\"", status = "200 OK", 
+        `content-type` = "application/gpx+xml; charset=utf-8", 
+        server = "Apache/2.4.54 (Ubuntu)"), class = "httr2_headers"), 
+    body = as.raw(c(0x3c, 0x3f, 0x78, 0x6d, 0x6c, 0x20, 0x76, 
+    0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x3d, 0x22, 0x31, 0x2e, 
+    0x30, 0x22, 0x20, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 
+    0x67, 0x3d, 0x22, 0x55, 0x54, 0x46, 0x2d, 0x38, 0x22, 0x3f, 
+    0x3e, 0x0a, 0x3c, 0x67, 0x70, 0x78, 0x20, 0x76, 0x65, 0x72, 
+    0x73, 0x69, 0x6f, 0x6e, 0x3d, 0x22, 0x31, 0x2e, 0x31, 0x22, 
[…]
