+++ tests/testthat/_snaps/elements.md
@@ -3,22 +3,24 @@
     Code
       print(x)
     Output
-        type       id visible version changeset           timestamp   user     uid
-      1 node 35308286    TRUE      15  63949815 2018-10-28 13:07:37 ManelG 2491053
+        type       id visible version changeset           timestamp     user      uid
+      1 node 35308286    TRUE      17 140341361 2023-08-24 20:19:22 jmaspons 11725140
                lat       lon members
       1 42.5189047 2.4565596    NULL
                                                                                        tags
-      1 5 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
+      1 7 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
 
 ---
 
     Code
       print(x)
     Output
-        type       id visible version changeset           timestamp     user      uid
-      1  way 13073736    TRUE      21 114656477 2021-12-07 11:27:19 jmaspons 11725140
-         lat  lon                                                         members
-      1 <NA> <NA> 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
+        type       id visible version changeset           timestamp
+      1  way 13073736    TRUE      27 140205821 2023-08-22 01:47:19
+                      user     uid  lat  lon
+      1 SomeoneElse_Revert 1778799 <NA> <NA>
+                                                                members
+      1 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
                                                                                        tags
       1 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
 
@@ -27,50 +29,56 @@
     Code
       print(x)
     Output
-            type    id visible version changeset           timestamp  user     uid
-      1 relation 40581    TRUE      53 135982205 2023-05-11 15:10:18 l2212 1775455
-         lat  lon                                                         members
-      1 <NA> <NA> 20 members: node/72994199/admin_centre, way/165579897/outer,...
+            type    id visible version changeset           timestamp        user
+      1 relation 40581    TRUE      62 151100114 2024-05-09 13:23:23 Siciliapars
+             uid  lat  lon
+      1 21349271 <NA> <NA>
+                                                                members
+      1 20 members: node/72994199/admin_centre, way/165579897/outer,...
                                                                                        tags
-      1 22 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
+      1 22 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
 
 # osm_history_object works
 
     Code
       print(x)
     Output
-         type       id visible version changeset           timestamp     user     uid
-      1  node 35308286    TRUE       1    292412 2007-09-02 14:55:33  Skywave   10927
-      2  node 35308286    TRUE       2     39997 2007-12-02 19:28:54  Skywave   10927
-      3  node 35308286    TRUE       3     39997 2007-12-02 19:29:01  Skywave   10927
-      4  node 35308286    TRUE       4    131335 2007-12-06 21:08:56  Skywave   10927
[…]

[…]
