### 2024-07-01 01618f5
Fix missing object in osm_get_objects(...,  full_object  = TRUE, format = "R")
FIX #35
