+++ tests/testthat/_snaps/elements.md
@@ -667,129 +667,129 @@
       print(x)
     Output
          type         id visible version changeset           timestamp
-      1  node 4856486143    TRUE       2  86534163 2020-06-11 21:02:34
-      2  node 4902243087    TRUE       3  76811325 2019-11-08 13:10:11
-      3  node 4902243088    TRUE       3  76811325 2019-11-08 13:10:11
-      4  node 6771540792    TRUE       2  77447443 2019-11-22 22:35:38
-      5  node 6771540793    TRUE       2  77447443 2019-11-22 22:35:38
-      6  node 6771540804    TRUE       2  76811325 2019-11-08 13:10:11
-      7  node 6771540805    TRUE       2  76811325 2019-11-08 13:10:11
-      8  node 6771540806    TRUE       2  76811325 2019-11-08 13:10:11
-      9  node 6771540812    TRUE       2  76811325 2019-11-08 13:10:11
-      10 node 6790061967    TRUE       2  76811325 2019-11-08 13:10:11
-      11 node 6957604940    TRUE       1  76811325 2019-11-08 13:10:11
-      12 node 6957604941    TRUE       1  76811325 2019-11-08 13:10:11
-      13 node 6957604951    TRUE       1  76811325 2019-11-08 13:10:11
-      14 node 6957604952    TRUE       2  77447443 2019-11-22 22:35:38
-      15 node 6957604954    TRUE       1  76811325 2019-11-08 13:10:11
-      16 node 6957604955    TRUE       1  76811325 2019-11-08 13:10:11
-      17 node 6957604959    TRUE       1  76811325 2019-11-08 13:10:11
-      18 node 6957604960    TRUE       1  76811325 2019-11-08 13:10:11
-      19 node 6957604961    TRUE       1  76811325 2019-11-08 13:10:11
-      20 node 6957604962    TRUE       1  76811325 2019-11-08 13:10:11
-      21 node 6957604963    TRUE       1  76811325 2019-11-08 13:10:11
-      22 node 6957604964    TRUE       1  76811325 2019-11-08 13:10:11
-      23 node 6957604965    TRUE       1  76811325 2019-11-08 13:10:11
-      24 node 6957604966    TRUE       1  76811325 2019-11-08 13:10:11
-      25 node 6957604967    TRUE       1  76811325 2019-11-08 13:10:11
-      26 node 6957604968    TRUE       1  76811325 2019-11-08 13:10:11
-      27 node 6957604969    TRUE       1  76811325 2019-11-08 13:10:11
-      28 node 6957604972    TRUE       1  76811325 2019-11-08 13:10:11
-      29 node 6957604973    TRUE       1  76811325 2019-11-08 13:10:11
-      30 node 6957604974    TRUE       1  76811325 2019-11-08 13:10:11
-      31 node 6957604975    TRUE       1  76811325 2019-11-08 13:10:11
-      32 node 6957604977    TRUE       1  76811325 2019-11-08 13:10:11
-      33 node 6957604979    TRUE       1  76811325 2019-11-08 13:10:11
-      34 node 6957604980    TRUE       1  76811325 2019-11-08 13:10:11
-      35 node 6957604981    TRUE       1  76811325 2019-11-08 13:10:11
-      36 node 6957604982    TRUE       1  76811325 2019-11-08 13:10:11
-      37 node 6957604983    TRUE       1  76811325 2019-11-08 13:10:11
-      38 node 6957604984    TRUE       1  76811325 2019-11-08 13:10:11
-      39 node 6957611385    TRUE       1  76811325 2019-11-08 13:10:11
-      40 node 6957611387    TRUE       1  76811325 2019-11-08 13:10:11
-      41 node 6957611388    TRUE       1  76811325 2019-11-08 13:10:11
[…]
