### 2024-07-01 78513b3
Increase version and add NEWS
