### 2024-07-01 82a845d
Deploying to gh-pages from @ jmaspons/osmapiR@78513b3d1c5465983469f72d4af03a2d8040aa4d 🚀
