### 2024-07-03 #633 @jmaspons
> Logged author response!

Is this enough, @adamhsparks ?
