### 2024-07-03 #633 @ropensci-review-bot
Approved! Thanks @jmaspons for submitting and @jonthegeek, @ccamara for your reviews! :grin:
